@@ -16,22 +16,24 @@
 namespace FastyBird\Connector\NsPanel\Commands;
 
 use DateTimeInterface;
-use FastyBird\Connector\NsPanel\Entities;
+use FastyBird\Connector\NsPanel\Documents;
 use FastyBird\Connector\NsPanel\Exceptions;
 use FastyBird\Connector\NsPanel\Helpers;
+use FastyBird\Connector\NsPanel\Queries;
 use FastyBird\DateTimeFactory;
-use FastyBird\Library\Metadata\Documents as MetadataDocuments;
 use FastyBird\Library\Metadata\Exceptions as MetadataExceptions;
 use FastyBird\Module\Devices\Commands as DevicesCommands;
 use FastyBird\Module\Devices\Exceptions as DevicesExceptions;
 use FastyBird\Module\Devices\Models as DevicesModels;
-use FastyBird\Module\Devices\Queries as DevicesQueries;
+use FastyBird\Module\Devices\Types as DevicesTypes;
 use Nette\Localization;
 use Ramsey\Uuid;
 use Symfony\Component\Console;
 use Symfony\Component\Console\Input;
 use Symfony\Component\Console\Output;
 use Symfony\Component\Console\Style;
+use TypeError;
+use ValueError;
 use function array_key_exists;
 use function array_key_first;
 use function array_search;
@@ -101,8 +103,11 @@ class Discover extends Console\Command\Command
 	 * @throws Console\Exception\ExceptionInterface
 	 * @throws Console\Exception\InvalidArgumentException
 	 * @throws DevicesExceptions\InvalidState
+	 * @throws Exceptions\InvalidArgument
 	 * @throws MetadataExceptions\InvalidArgument
 	 * @throws MetadataExceptions\InvalidState
+	 * @throws TypeError
+	 * @throws ValueError
 	 */
 	protected function execute(Input\InputInterface $input, Output\OutputInterface $output): int
 	{
@@ -138,8 +143,7 @@ class Discover extends Console\Command\Command
 		) {
 			$connectorId = $input->getOption('connector');
 
-			$findConnectorQuery = new DevicesQueries\Configuration\FindConnectors();
-			$findConnectorQuery->byType(Entities\NsPanelChannel::TYPE);
+			$findConnectorQuery = new Queries\Configuration\FindConnectors();
 
 			if (Uuid\Uuid::isValid($connectorId)) {
 				$findConnectorQuery->byId(Uuid\Uuid::fromString($connectorId));
@@ -147,7 +151,10 @@ class Discover extends Console\Command\Command
 				$findConnectorQuery->byIdentifier($connectorId);
 			}
 
-			$connector = $this->connectorsConfigurationRepository->findOneBy($findConnectorQuery);
+			$connector = $this->connectorsConfigurationRepository->findOneBy(
+				$findConnectorQuery,
+				Documents\Connectors\Connector::class,
+			);
 
 			if ($connector === null) {
 				$io->warning(
@@ -159,14 +166,15 @@ class Discover extends Console\Command\Command
 		} else {
 			$connectors = [];
 
-			$findConnectorsQuery = new DevicesQueries\Configuration\FindConnectors();
-			$findConnectorsQuery->byType(Entities\NsPanelChannel::TYPE);
+			$findConnectorsQuery = new Queries\Configuration\FindConnectors();
 
-			$systemConnectors = $this->connectorsConfigurationRepository->findAllBy($findConnectorsQuery);
+			$systemConnectors = $this->connectorsConfigurationRepository->findAllBy(
+				$findConnectorsQuery,
+				Documents\Connectors\Connector::class,
+			);
 			usort(
 				$systemConnectors,
-				// phpcs:ignore SlevomatCodingStandard.Files.LineLength.LineTooLong
-				static fn (MetadataDocuments\DevicesModule\Connector $a, MetadataDocuments\DevicesModule\Connector $b): int => $a->getIdentifier() <=> $b->getIdentifier()
+				static fn (Documents\Connectors\Connector $a, Documents\Connectors\Connector $b): int => $a->getIdentifier() <=> $b->getIdentifier()
 			);
 
 			foreach ($systemConnectors as $connector) {
@@ -183,11 +191,13 @@ class Discover extends Console\Command\Command
 			if (count($connectors) === 1) {
 				$connectorIdentifier = array_key_first($connectors);
 
-				$findConnectorQuery = new DevicesQueries\Configuration\FindConnectors();
+				$findConnectorQuery = new Queries\Configuration\FindConnectors();
 				$findConnectorQuery->byIdentifier($connectorIdentifier);
-				$findConnectorQuery->byType(Entities\NsPanelChannel::TYPE);
 
-				$connector = $this->connectorsConfigurationRepository->findOneBy($findConnectorQuery);
+				$connector = $this->connectorsConfigurationRepository->findOneBy(
+					$findConnectorQuery,
+					Documents\Connectors\Connector::class,
+				);
 
 				if ($connector === null) {
 					$io->warning(
@@ -219,7 +229,7 @@ class Discover extends Console\Command\Command
 					$this->translator->translate('//ns-panel-connector.cmd.base.messages.answerNotValid'),
 				);
 				$question->setValidator(
-					function (string|int|null $answer) use ($connectors): MetadataDocuments\DevicesModule\Connector {
+					function (string|int|null $answer) use ($connectors): Documents\Connectors\Connector {
 						if ($answer === null) {
 							throw new Exceptions\Runtime(
 								sprintf(
@@ -238,11 +248,13 @@ class Discover extends Console\Command\Command
 						$identifier = array_search($answer, $connectors, true);
 
 						if ($identifier !== false) {
-							$findConnectorQuery = new DevicesQueries\Configuration\FindConnectors();
+							$findConnectorQuery = new Queries\Configuration\FindConnectors();
 							$findConnectorQuery->byIdentifier($identifier);
-							$findConnectorQuery->byType(Entities\NsPanelChannel::TYPE);
 
-							$connector = $this->connectorsConfigurationRepository->findOneBy($findConnectorQuery);
+							$connector = $this->connectorsConfigurationRepository->findOneBy(
+								$findConnectorQuery,
+								Documents\Connectors\Connector::class,
+							);
 
 							if ($connector !== null) {
 								return $connector;
@@ -259,7 +271,7 @@ class Discover extends Console\Command\Command
 				);
 
 				$connector = $io->askQuestion($question);
-				assert($connector instanceof MetadataDocuments\DevicesModule\Connector);
+				assert($connector instanceof Documents\Connectors\Connector);
 			}
 		}
 
@@ -279,7 +291,7 @@ class Discover extends Console\Command\Command
 
 		$result = $serviceCmd->run(new Input\ArrayInput([
 			'--connector' => $connector->getId()->toString(),
-			'--mode' => DevicesCommands\Connector::MODE_DISCOVER,
+			'--mode' => DevicesTypes\ConnectorMode::DISCOVER->value,
 			'--no-interaction' => true,
 			'--quiet' => true,
 		]), $output);
@@ -301,13 +313,16 @@ class Discover extends Console\Command\Command
 
 	/**
 	 * @throws DevicesExceptions\InvalidState
+	 * @throws Exceptions\InvalidArgument
 	 * @throws MetadataExceptions\InvalidArgument
 	 * @throws MetadataExceptions\InvalidState
+	 * @throws TypeError
+	 * @throws ValueError
 	 */
 	private function showResults(
 		Style\SymfonyStyle $io,
 		Output\OutputInterface $output,
-		MetadataDocuments\DevicesModule\Connector $connector,
+		Documents\Connectors\Connector $connector,
 	): void
 	{
 		$table = new Console\Helper\Table($output);
@@ -321,19 +336,23 @@ class Discover extends Console\Command\Command
 
 		$foundDevices = 0;
 
-		$findDevicesQuery = new DevicesQueries\Configuration\FindDevices();
+		$findDevicesQuery = new Queries\Configuration\FindGatewayDevices();
 		$findDevicesQuery->forConnector($connector);
-		$findDevicesQuery->byType(Entities\Devices\Gateway::TYPE);
 
-		$gateways = $this->devicesConfigurationRepository->findAllBy($findDevicesQuery);
+		$gateways = $this->devicesConfigurationRepository->findAllBy(
+			$findDevicesQuery,
+			Documents\Devices\Gateway::class,
+		);
 
 		foreach ($gateways as $gateway) {
-			$findDevicesQuery = new DevicesQueries\Configuration\FindDevices();
+			$findDevicesQuery = new Queries\Configuration\FindSubDevices();
 			$findDevicesQuery->forConnector($connector);
 			$findDevicesQuery->forParent($gateway);
-			$findDevicesQuery->byType(Entities\Devices\SubDevice::TYPE);
 
-			$devices = $this->devicesConfigurationRepository->findAllBy($findDevicesQuery);
+			$devices = $this->devicesConfigurationRepository->findAllBy(
+				$findDevicesQuery,
+				Documents\Devices\SubDevice::class,
+			);
 
 			foreach ($devices as $device) {
 				$createdAt = $device->getCreatedAt();
