@@ -15,14 +15,13 @@
 
 namespace FastyBird\Connector\NsPanel\Commands;
 
-use FastyBird\Connector\NsPanel\Entities;
+use FastyBird\Connector\NsPanel\Documents;
 use FastyBird\Connector\NsPanel\Exceptions;
-use FastyBird\Library\Metadata\Documents as MetadataDocuments;
+use FastyBird\Connector\NsPanel\Queries;
 use FastyBird\Library\Metadata\Exceptions as MetadataExceptions;
 use FastyBird\Module\Devices\Commands as DevicesCommands;
 use FastyBird\Module\Devices\Exceptions as DevicesExceptions;
 use FastyBird\Module\Devices\Models as DevicesModels;
-use FastyBird\Module\Devices\Queries as DevicesQueries;
 use Nette\Localization;
 use Ramsey\Uuid;
 use Symfony\Component\Console;
@@ -127,8 +126,7 @@ class Execute extends Console\Command\Command
 		) {
 			$connectorId = $input->getOption('connector');
 
-			$findConnectorQuery = new DevicesQueries\Configuration\FindConnectors();
-			$findConnectorQuery->byType(Entities\NsPanelConnector::TYPE);
+			$findConnectorQuery = new Queries\Configuration\FindConnectors();
 
 			if (Uuid\Uuid::isValid($connectorId)) {
 				$findConnectorQuery->byId(Uuid\Uuid::fromString($connectorId));
@@ -136,7 +134,10 @@ class Execute extends Console\Command\Command
 				$findConnectorQuery->byIdentifier($connectorId);
 			}
 
-			$connector = $this->connectorsConfigurationRepository->findOneBy($findConnectorQuery);
+			$connector = $this->connectorsConfigurationRepository->findOneBy(
+				$findConnectorQuery,
+				Documents\Connectors\Connector::class,
+			);
 
 			if ($connector === null) {
 				$io->warning(
@@ -148,14 +149,15 @@ class Execute extends Console\Command\Command
 		} else {
 			$connectors = [];
 
-			$findConnectorsQuery = new DevicesQueries\Configuration\FindConnectors();
-			$findConnectorsQuery->byType(Entities\NsPanelConnector::TYPE);
+			$findConnectorsQuery = new Queries\Configuration\FindConnectors();
 
-			$systemConnectors = $this->connectorsConfigurationRepository->findAllBy($findConnectorsQuery);
+			$systemConnectors = $this->connectorsConfigurationRepository->findAllBy(
+				$findConnectorsQuery,
+				Documents\Connectors\Connector::class,
+			);
 			usort(
 				$systemConnectors,
-				// phpcs:ignore SlevomatCodingStandard.Files.LineLength.LineTooLong
-				static fn (MetadataDocuments\DevicesModule\Connector $a, MetadataDocuments\DevicesModule\Connector $b): int => $a->getIdentifier() <=> $b->getIdentifier()
+				static fn (Documents\Connectors\Connector $a, Documents\Connectors\Connector $b): int => $a->getIdentifier() <=> $b->getIdentifier()
 			);
 
 			foreach ($systemConnectors as $connector) {
@@ -172,11 +174,13 @@ class Execute extends Console\Command\Command
 			if (count($connectors) === 1) {
 				$connectorIdentifier = array_key_first($connectors);
 
-				$findConnectorQuery = new DevicesQueries\Configuration\FindConnectors();
+				$findConnectorQuery = new Queries\Configuration\FindConnectors();
 				$findConnectorQuery->byIdentifier($connectorIdentifier);
-				$findConnectorQuery->byType(Entities\NsPanelConnector::TYPE);
 
-				$connector = $this->connectorsConfigurationRepository->findOneBy($findConnectorQuery);
+				$connector = $this->connectorsConfigurationRepository->findOneBy(
+					$findConnectorQuery,
+					Documents\Connectors\Connector::class,
+				);
 
 				if ($connector === null) {
 					$io->warning(
@@ -208,7 +212,7 @@ class Execute extends Console\Command\Command
 					$this->translator->translate('//ns-panel-connector.cmd.base.messages.answerNotValid'),
 				);
 				$question->setValidator(
-					function (string|int|null $answer) use ($connectors): MetadataDocuments\DevicesModule\Connector {
+					function (string|int|null $answer) use ($connectors): Documents\Connectors\Connector {
 						if ($answer === null) {
 							throw new Exceptions\Runtime(
 								sprintf(
@@ -227,11 +231,13 @@ class Execute extends Console\Command\Command
 						$identifier = array_search($answer, $connectors, true);
 
 						if ($identifier !== false) {
-							$findConnectorQuery = new DevicesQueries\Configuration\FindConnectors();
+							$findConnectorQuery = new Queries\Configuration\FindConnectors();
 							$findConnectorQuery->byIdentifier($identifier);
-							$findConnectorQuery->byType(Entities\NsPanelConnector::TYPE);
 
-							$connector = $this->connectorsConfigurationRepository->findOneBy($findConnectorQuery);
+							$connector = $this->connectorsConfigurationRepository->findOneBy(
+								$findConnectorQuery,
+								Documents\Connectors\Connector::class,
+							);
 
 							if ($connector !== null) {
 								return $connector;
@@ -248,7 +254,7 @@ class Execute extends Console\Command\Command
 				);
 
 				$connector = $io->askQuestion($question);
-				assert($connector instanceof MetadataDocuments\DevicesModule\Connector);
+				assert($connector instanceof Documents\Connectors\Connector);
 			}
 		}
 
