@@ -17,25 +17,27 @@ namespace FastyBird\Connector\NsPanel\Connector;
 
 use FastyBird\Connector\NsPanel;
 use FastyBird\Connector\NsPanel\Clients;
-use FastyBird\Connector\NsPanel\Entities;
+use FastyBird\Connector\NsPanel\Documents;
 use FastyBird\Connector\NsPanel\Exceptions;
 use FastyBird\Connector\NsPanel\Helpers;
+use FastyBird\Connector\NsPanel\Queries;
 use FastyBird\Connector\NsPanel\Queue;
 use FastyBird\Connector\NsPanel\Servers;
 use FastyBird\Connector\NsPanel\Writers;
-use FastyBird\Library\Bootstrap\Helpers as BootstrapHelpers;
-use FastyBird\Library\Metadata\Documents as MetadataDocuments;
+use FastyBird\Library\Application\Helpers as ApplicationHelpers;
+use FastyBird\Library\Exchange\Exceptions as ExchangeExceptions;
 use FastyBird\Library\Metadata\Exceptions as MetadataExceptions;
 use FastyBird\Library\Metadata\Types as MetadataTypes;
 use FastyBird\Module\Devices\Connectors as DevicesConnectors;
-use FastyBird\Module\Devices\Events as DevicesEvents;
+use FastyBird\Module\Devices\Documents as DevicesDocuments;
 use FastyBird\Module\Devices\Exceptions as DevicesExceptions;
 use FastyBird\Module\Devices\Models as DevicesModels;
-use FastyBird\Module\Devices\Queries as DevicesQueries;
 use Nette;
-use Psr\EventDispatcher as PsrEventDispatcher;
 use React\EventLoop;
+use React\Promise;
 use ReflectionClass;
+use TypeError;
+use ValueError;
 use function array_key_exists;
 use function assert;
 use function React\Async\async;
@@ -66,41 +68,48 @@ final class Connector implements DevicesConnectors\Connector
 
 	/**
 	 * @param array<Clients\ClientFactory> $clientsFactories
+	 * @param array<Writers\WriterFactory> $writersFactories
 	 */
 	public function __construct(
-		private readonly MetadataDocuments\DevicesModule\Connector $connector,
+		private readonly DevicesDocuments\Connectors\Connector $connector,
 		private readonly array $clientsFactories,
 		private readonly Clients\DiscoveryFactory $discoveryClientFactory,
-		private readonly Helpers\Entity $entityHelper,
-		private readonly Helpers\Connector $connectorHelper,
+		private readonly Helpers\MessageBuilder $messageBuilder,
+		private readonly Helpers\Connectors\Connector $connectorHelper,
 		private readonly Helpers\Devices\ThirdPartyDevice $thirdPartyDeviceHelper,
 		private readonly Queue\Queue $queue,
 		private readonly Queue\Consumers $consumers,
 		private readonly Servers\ServerFactory $serverFactory,
-		private readonly Writers\WriterFactory $writerFactory,
+		private readonly array $writersFactories,
 		private readonly NsPanel\Logger $logger,
 		private readonly DevicesModels\Configuration\Devices\Repository $devicesConfigurationRepository,
 		private readonly DevicesModels\Configuration\Channels\Repository $channelsConfigurationRepository,
 		private readonly EventLoop\LoopInterface $eventLoop,
-		private readonly PsrEventDispatcher\EventDispatcherInterface|null $dispatcher = null,
 	)
 	{
+		assert($this->connector instanceof Documents\Connectors\Connector);
 	}
 
 	/**
+	 * @return Promise\PromiseInterface<bool>
+	 *
 	 * @throws DevicesExceptions\InvalidState
+	 * @throws Exceptions\InvalidArgument
 	 * @throws Exceptions\InvalidState
+	 * @throws ExchangeExceptions\InvalidArgument
 	 * @throws MetadataExceptions\InvalidArgument
 	 * @throws MetadataExceptions\InvalidState
+	 * @throws TypeError
+	 * @throws ValueError
 	 */
-	public function execute(): void
+	public function execute(bool $standalone = true): Promise\PromiseInterface
 	{
-		assert($this->connector->getType() === Entities\NsPanelConnector::TYPE);
+		assert($this->connector instanceof Documents\Connectors\Connector);
 
 		$this->logger->info(
 			'Starting NS Panel connector service',
 			[
-				'source' => MetadataTypes\ConnectorSource::SOURCE_CONNECTOR_NS_PANEL,
+				'source' => MetadataTypes\Sources\Connector::NS_PANEL->value,
 				'type' => 'connector',
 				'connector' => [
 					'id' => $this->connector->getId()->toString(),
@@ -118,8 +127,8 @@ final class Connector implements DevicesConnectors\Connector
 			if (
 				(
 					array_key_exists(Clients\ClientFactory::MODE_CONSTANT_NAME, $constants)
-					&& $mode->equalsValue($constants[Clients\ClientFactory::MODE_CONSTANT_NAME])
-				) || $mode->equalsValue(NsPanel\Types\ClientMode::BOTH)
+					&& $mode === $constants[Clients\ClientFactory::MODE_CONSTANT_NAME]
+				) || $mode === NsPanel\Types\ClientMode::BOTH
 			) {
 				$client = $clientFactory->create($this->connector);
 				$client->connect();
@@ -129,15 +138,27 @@ final class Connector implements DevicesConnectors\Connector
 		}
 
 		if (
-			$mode->equalsValue(NsPanel\Types\ClientMode::BOTH)
-			|| $mode->equalsValue(NsPanel\Types\ClientMode::DEVICE)
+			$mode === NsPanel\Types\ClientMode::BOTH
+			|| $mode === NsPanel\Types\ClientMode::DEVICE
 		) {
 			$this->server = $this->serverFactory->create($this->connector);
 			$this->server->connect();
 		}
 
-		$this->writer = $this->writerFactory->create($this->connector);
-		$this->writer->connect();
+		foreach ($this->writersFactories as $writerFactory) {
+			if (
+				(
+					$standalone
+					&& $writerFactory instanceof Writers\ExchangeFactory
+				) || (
+					!$standalone
+					&& $writerFactory instanceof Writers\EventFactory
+				)
+			) {
+				$this->writer = $writerFactory->create($this->connector);
+				$this->writer->connect();
+			}
+		}
 
 		$this->consumersTimer = $this->eventLoop->addPeriodicTimer(
 			self::QUEUE_PROCESSING_INTERVAL,
@@ -147,27 +168,32 @@ final class Connector implements DevicesConnectors\Connector
 		);
 
 		if (
-			$mode->equalsValue(NsPanel\Types\ClientMode::BOTH)
-			|| $mode->equalsValue(NsPanel\Types\ClientMode::DEVICE)
+			$mode === NsPanel\Types\ClientMode::BOTH
+			|| $mode === NsPanel\Types\ClientMode::DEVICE
 		) {
-			$this->eventLoop->addTimer(1, function (): void {
-				$findDevicesQuery = new DevicesQueries\Configuration\FindDevices();
+			$this->eventLoop->addTimer(1, async(function (): void {
+				$findDevicesQuery = new Queries\Configuration\FindThirdPartyDevices();
 				$findDevicesQuery->forConnector($this->connector);
-				$findDevicesQuery->byType(Entities\Devices\ThirdPartyDevice::TYPE);
 
-				$devices = $this->devicesConfigurationRepository->findAllBy($findDevicesQuery);
+				$devices = $this->devicesConfigurationRepository->findAllBy(
+					$findDevicesQuery,
+					Documents\Devices\ThirdPartyDevice::class,
+				);
 
 				foreach ($devices as $device) {
-					$findChannelsQuery = new DevicesQueries\Configuration\FindChannels();
+					$findChannelsQuery = new Queries\Configuration\FindChannels();
 					$findChannelsQuery->forDevice($device);
 
-					$channels = $this->channelsConfigurationRepository->findAllBy($findChannelsQuery);
+					$channels = $this->channelsConfigurationRepository->findAllBy(
+						$findChannelsQuery,
+						Documents\Channels\Channel::class,
+					);
 
 					foreach ($channels as $channel) {
 						try {
 							$this->queue->append(
-								$this->entityHelper->create(
-									Entities\Messages\WriteThirdPartyDeviceState::class,
+								$this->messageBuilder->create(
+									Queue\Messages\WriteThirdPartyDeviceState::class,
 									[
 										'connector' => $device->getConnector(),
 										'device' => $device->getId(),
@@ -177,11 +203,11 @@ final class Connector implements DevicesConnectors\Connector
 							);
 						} catch (Exceptions\Runtime $ex) {
 							$this->logger->error(
-								'Could report device initial state to NS Panel',
+								'Could not report device initial state to NS Panel',
 								[
-									'source' => MetadataTypes\ConnectorSource::SOURCE_CONNECTOR_NS_PANEL,
+									'source' => MetadataTypes\Sources\Connector::NS_PANEL->value,
 									'type' => 'connector',
-									'exception' => BootstrapHelpers\Logger::buildException($ex),
+									'exception' => ApplicationHelpers\Logger::buildException($ex),
 									'connector' => [
 										'id' => $this->connector->getId()->toString(),
 									],
@@ -199,34 +225,41 @@ final class Connector implements DevicesConnectors\Connector
 						}
 					}
 				}
-			});
+			}));
 		}
 
 		$this->logger->info(
 			'NS Panel connector service has been started',
 			[
-				'source' => MetadataTypes\ConnectorSource::SOURCE_CONNECTOR_NS_PANEL,
+				'source' => MetadataTypes\Sources\Connector::NS_PANEL->value,
 				'type' => 'connector',
 				'connector' => [
 					'id' => $this->connector->getId()->toString(),
 				],
 			],
 		);
+
+		return Promise\resolve(true);
 	}
 
 	/**
+	 * @return Promise\PromiseInterface<bool>
+	 *
 	 * @throws DevicesExceptions\InvalidState
+	 * @throws Exceptions\InvalidArgument
 	 * @throws MetadataExceptions\InvalidArgument
 	 * @throws MetadataExceptions\InvalidState
+	 * @throws TypeError
+	 * @throws ValueError
 	 */
-	public function discover(): void
+	public function discover(): Promise\PromiseInterface
 	{
-		assert($this->connector->getType() === Entities\NsPanelConnector::TYPE);
+		assert($this->connector instanceof Documents\Connectors\Connector);
 
 		$this->logger->info(
 			'Starting NS Panel connector discovery',
 			[
-				'source' => MetadataTypes\ConnectorSource::SOURCE_CONNECTOR_NS_PANEL,
+				'source' => MetadataTypes\Sources\Connector::NS_PANEL->value,
 				'type' => 'connector',
 				'connector' => [
 					'id' => $this->connector->getId()->toString(),
@@ -235,15 +268,6 @@ final class Connector implements DevicesConnectors\Connector
 		);
 
 		$client = $this->discoveryClientFactory->create($this->connector);
-
-		$client->on('finished', function (): void {
-			$this->dispatcher?->dispatch(
-				new DevicesEvents\TerminateConnector(
-					MetadataTypes\ConnectorSource::get(MetadataTypes\ConnectorSource::SOURCE_CONNECTOR_NS_PANEL),
-					'Devices discovery finished',
-				),
-			);
-		});
 
 		$this->clients[] = $client;
 
@@ -257,7 +281,7 @@ final class Connector implements DevicesConnectors\Connector
 		$this->logger->info(
 			'NS Panel connector discovery has been started',
 			[
-				'source' => MetadataTypes\ConnectorSource::SOURCE_CONNECTOR_NS_PANEL,
+				'source' => MetadataTypes\Sources\Connector::NS_PANEL->value,
 				'type' => 'connector',
 				'connector' => [
 					'id' => $this->connector->getId()->toString(),
@@ -266,11 +290,13 @@ final class Connector implements DevicesConnectors\Connector
 		);
 
 		$client->discover();
+
+		return Promise\resolve(true);
 	}
 
 	public function terminate(): void
 	{
-		assert($this->connector->getType() === Entities\NsPanelConnector::TYPE);
+		assert($this->connector instanceof Documents\Connectors\Connector);
 
 		foreach ($this->clients as $client) {
 			$client->disconnect();
@@ -287,7 +313,7 @@ final class Connector implements DevicesConnectors\Connector
 		$this->logger->info(
 			'NS Panel connector has been terminated',
 			[
-				'source' => MetadataTypes\ConnectorSource::SOURCE_CONNECTOR_NS_PANEL,
+				'source' => MetadataTypes\Sources\Connector::NS_PANEL->value,
 				'type' => 'connector',
 				'connector' => [
 					'id' => $this->connector->getId()->toString(),
