@@ -17,6 +17,8 @@ namespace FastyBird\Connector\NsPanel\Helpers;
 
 use FastyBird\Connector\NsPanel\Types;
 use Nette\Utils;
+use TypeError;
+use ValueError;
 use function lcfirst;
 use function preg_replace;
 use function str_replace;
@@ -40,7 +42,7 @@ final class Name
 		string|int|null $name = null,
 	): string
 	{
-		$identifier = str_replace('-', '_', Utils\Strings::lower($capability->getValue()));
+		$identifier = str_replace('-', '_', Utils\Strings::lower($capability->value));
 
 		if ($name !== null) {
 			$identifier .= '_' . $name;
@@ -51,12 +53,16 @@ final class Name
 
 	public static function convertProtocolToProperty(Types\Protocol $protocol): string
 	{
-		return strtolower(strval(preg_replace('/(?<!^)[A-Z]/', '_$0', $protocol->getValue())));
+		return strtolower(strval(preg_replace('/(?<!^)[A-Z]/', '_$0', $protocol->value)));
 	}
 
+	/**
+	 * @throws TypeError
+	 * @throws ValueError
+	 */
 	public static function convertPropertyToProtocol(string $identifier): Types\Protocol
 	{
-		return Types\Protocol::get(
+		return Types\Protocol::from(
 			lcfirst(
 				str_replace(
 					' ',
