@@ -15,24 +15,98 @@
 
 namespace FastyBird\Connector\NsPanel\Hydrators\Devices;
 
+use Doctrine\Common;
+use Doctrine\Persistence;
 use FastyBird\Connector\NsPanel\Entities;
-use FastyBird\Connector\NsPanel\Hydrators;
+use FastyBird\Connector\NsPanel\Schemas;
+use FastyBird\JsonApi\Exceptions as JsonApiExceptions;
+use FastyBird\JsonApi\Helpers;
+use FastyBird\Library\Application\Exceptions as ApplicationExceptions;
+use FastyBird\Module\Devices\Entities as DevicesEntities;
+use FastyBird\Module\Devices\Models as DevicesModels;
+use Fig\Http\Message\StatusCodeInterface;
+use IPub\JsonAPIDocument;
+use Nette\Localization;
+use Ramsey\Uuid;
+use function is_string;
 
 /**
  * NS Panel sub-device entity hydrator
  *
- * @extends Hydrators\NsPanelDevice<Entities\Devices\SubDevice>
+ * @extends Device<Entities\Devices\SubDevice>
  *
  * @package        FastyBird:NsPanelConnector!
  * @subpackage     Hydrators
  * @author         Adam Kadlec <adam.kadlec@fastybird.com>
  */
-final class SubDevice extends Hydrators\NsPanelDevice
+final class SubDevice extends Device
 {
+
+	public function __construct(
+		private readonly DevicesModels\Entities\Devices\DevicesRepository $devicesRepository,
+		DevicesModels\Entities\Connectors\ConnectorsRepository $connectorsRepository,
+		Persistence\ManagerRegistry $managerRegistry,
+		Localization\Translator $translator,
+		Helpers\CrudReader|null $crudReader = null,
+		Common\Cache\Cache|null $cache = null,
+	)
+	{
+		parent::__construct($connectorsRepository, $managerRegistry, $translator, $crudReader, $cache);
+	}
 
 	public function getEntityName(): string
 	{
 		return Entities\Devices\SubDevice::class;
+	}
+
+	/**
+	 * @return array<DevicesEntities\Devices\Device>
+	 *
+	 * @throws ApplicationExceptions\InvalidState
+	 * @throws JsonApiExceptions\JsonApiError
+	 */
+	protected function hydrateParentsRelationship(
+		JsonAPIDocument\Objects\IRelationshipObject $relationships,
+		JsonAPIDocument\Objects\IResourceObjectCollection|null $included,
+		Entities\Devices\Gateway|null $entity,
+	): array
+	{
+		if ($relationships->getData() instanceof JsonAPIDocument\Objects\ResourceIdentifierCollection) {
+			$parents = [];
+			$foundValidParent = false;
+
+			foreach ($relationships->getData() as $relationship) {
+				if (
+					is_string($relationship->getId())
+					&& Uuid\Uuid::isValid($relationship->getId())
+				) {
+					$parent = $this->devicesRepository->find(
+						Uuid\Uuid::fromString($relationship->getId()),
+					);
+
+					if ($parent instanceof Entities\Devices\Gateway) {
+						$foundValidParent = true;
+					}
+
+					if ($parent !== null) {
+						$parents[] = $parent;
+					}
+				}
+			}
+
+			if ($parents !== [] && $foundValidParent) {
+				return $parents;
+			}
+		}
+
+		throw new JsonApiExceptions\JsonApiError(
+			StatusCodeInterface::STATUS_UNPROCESSABLE_ENTITY,
+			$this->translator->translate('//ns-panel-connector.base.messages.invalidRelation.heading'),
+			$this->translator->translate('//ns-panel-connector.base.messages.invalidRelation.message'),
+			[
+				'pointer' => '/data/relationships/' . Schemas\Devices\SubDevice::RELATIONSHIPS_PARENTS . '/data/id',
+			],
+		);
 	}
 
 }
