@@ -21,7 +21,7 @@ use FastyBird\Module\Devices\Queries as DevicesQueries;
 /**
  * Find connectors entities query
  *
- * @template T of Entities\NsPanelConnector
+ * @template T of Entities\Connectors\Connector
  * @extends  DevicesQueries\Entities\FindConnectors<T>
  *
  * @package        FastyBird:NsPanelConnector!
