@@ -18,12 +18,13 @@ namespace FastyBird\Connector\NsPanel\Queue\Consumers;
 use Doctrine\DBAL;
 use FastyBird\Connector\NsPanel;
 use FastyBird\Connector\NsPanel\Entities;
+use FastyBird\Connector\NsPanel\Exceptions;
+use FastyBird\Connector\NsPanel\Queries;
+use FastyBird\Library\Application\Exceptions as ApplicationExceptions;
+use FastyBird\Library\Application\Helpers as ApplicationHelpers;
 use FastyBird\Library\Metadata\Types as MetadataTypes;
 use FastyBird\Module\Devices\Entities as DevicesEntities;
-use FastyBird\Module\Devices\Exceptions as DevicesExceptions;
 use FastyBird\Module\Devices\Models as DevicesModels;
-use FastyBird\Module\Devices\Queries as DevicesQueries;
-use FastyBird\Module\Devices\Utilities as DevicesUtilities;
 use Nette\Utils;
 use Ramsey\Uuid;
 
@@ -38,7 +39,7 @@ use Ramsey\Uuid;
  * @property-read DevicesModels\Entities\Devices\DevicesRepository $devicesRepository
  * @property-read DevicesModels\Entities\Devices\Properties\PropertiesRepository $devicesPropertiesRepository
  * @property-read DevicesModels\Entities\Devices\Properties\PropertiesManager $devicesPropertiesManager
- * @property-read DevicesUtilities\Database $databaseHelper
+ * @property-read ApplicationHelpers\Database $databaseHelper
  * @property-read NsPanel\Logger $logger
  */
 trait DeviceProperty
@@ -47,20 +48,21 @@ trait DeviceProperty
 	/**
 	 * @param string|array<int, string>|array<int, string|int|float|array<int, string|int|float>|Utils\ArrayHash|null>|array<int, array<int, string|array<int, string|int|float|bool>|Utils\ArrayHash|null>>|null $format
 	 *
+	 * @throws ApplicationExceptions\InvalidState
+	 * @throws ApplicationExceptions\Runtime
+	 * @throws Exceptions\InvalidArgument
 	 * @throws DBAL\Exception
-	 * @throws DevicesExceptions\InvalidState
-	 * @throws DevicesExceptions\Runtime
 	 */
 	private function setDeviceProperty(
 		Uuid\UuidInterface $deviceId,
 		string|bool|int|null $value,
 		MetadataTypes\DataType $dataType,
-		string $identifier,
+		NsPanel\Types\DevicePropertyIdentifier $identifier,
 		string|null $name = null,
 		array|string|null $format = null,
 	): void
 	{
-		$findDevicePropertyQuery = new DevicesQueries\Entities\FindDeviceProperties();
+		$findDevicePropertyQuery = new Queries\Entities\FindDeviceProperties();
 		$findDevicePropertyQuery->byDeviceId($deviceId);
 		$findDevicePropertyQuery->byIdentifier($identifier);
 
@@ -94,14 +96,14 @@ trait DeviceProperty
 				$this->logger->warning(
 					'Stored device property was not of valid type',
 					[
-						'source' => MetadataTypes\ConnectorSource::SOURCE_CONNECTOR_NS_PANEL,
+						'source' => MetadataTypes\Sources\Connector::NS_PANEL->value,
 						'type' => 'message-consumer',
 						'device' => [
 							'id' => $deviceId->toString(),
 						],
 						'property' => [
 							'id' => $property->getId()->toString(),
-							'identifier' => $identifier,
+							'identifier' => $identifier->value,
 						],
 					],
 				);
@@ -113,20 +115,20 @@ trait DeviceProperty
 		if ($property === null) {
 			$device = $this->devicesRepository->find(
 				$deviceId,
-				Entities\NsPanelDevice::class,
+				Entities\Devices\Device::class,
 			);
 
 			if ($device === null) {
 				$this->logger->error(
 					'Device was not found, property could not be configured',
 					[
-						'source' => MetadataTypes\ConnectorSource::SOURCE_CONNECTOR_NS_PANEL,
+						'source' => MetadataTypes\Sources\Connector::NS_PANEL->value,
 						'type' => 'message-consumer',
 						'device' => [
 							'id' => $deviceId->toString(),
 						],
 						'property' => [
-							'identifier' => $identifier,
+							'identifier' => $identifier->value,
 						],
 					],
 				);
@@ -139,7 +141,7 @@ trait DeviceProperty
 					Utils\ArrayHash::from([
 						'entity' => DevicesEntities\Devices\Properties\Variable::class,
 						'device' => $device,
-						'identifier' => $identifier,
+						'identifier' => $identifier->value,
 						'name' => $name,
 						'dataType' => $dataType,
 						'value' => $value,
@@ -151,14 +153,14 @@ trait DeviceProperty
 			$this->logger->debug(
 				'Device variable property was created',
 				[
-					'source' => MetadataTypes\ConnectorSource::SOURCE_CONNECTOR_NS_PANEL,
+					'source' => MetadataTypes\Sources\Connector::NS_PANEL->value,
 					'type' => 'message-consumer',
 					'device' => [
 						'id' => $deviceId->toString(),
 					],
 					'property' => [
 						'id' => $property->getId()->toString(),
-						'identifier' => $identifier,
+						'identifier' => $identifier->value,
 					],
 				],
 			);
@@ -178,14 +180,14 @@ trait DeviceProperty
 			$this->logger->debug(
 				'Device variable property was updated',
 				[
-					'source' => MetadataTypes\ConnectorSource::SOURCE_CONNECTOR_NS_PANEL,
+					'source' => MetadataTypes\Sources\Connector::NS_PANEL->value,
 					'type' => 'message-consumer',
 					'device' => [
 						'id' => $deviceId->toString(),
 					],
 					'property' => [
 						'id' => $property->getId()->toString(),
-						'identifier' => $identifier,
+						'identifier' => $identifier->value,
 					],
 				],
 			);
