@@ -17,16 +17,23 @@ namespace FastyBird\Connector\NsPanel\Queue\Consumers;
 
 use Doctrine\DBAL;
 use FastyBird\Connector\NsPanel;
-use FastyBird\Connector\NsPanel\Entities;
+use FastyBird\Connector\NsPanel\Documents;
+use FastyBird\Connector\NsPanel\Queries;
 use FastyBird\Connector\NsPanel\Queue;
-use FastyBird\Library\Metadata\Documents as MetadataDocuments;
+use FastyBird\Library\Application\Exceptions as ApplicationExceptions;
 use FastyBird\Library\Metadata\Exceptions as MetadataExceptions;
 use FastyBird\Library\Metadata\Types as MetadataTypes;
+use FastyBird\Library\Tools\Exceptions as ToolsExceptions;
+use FastyBird\Module\Devices\Documents as DevicesDocuments;
 use FastyBird\Module\Devices\Exceptions as DevicesExceptions;
 use FastyBird\Module\Devices\Models as DevicesModels;
 use FastyBird\Module\Devices\Queries as DevicesQueries;
+use FastyBird\Module\Devices\Types as DevicesTypes;
 use FastyBird\Module\Devices\Utilities as DevicesUtilities;
 use Nette;
+use TypeError;
+use ValueError;
+use function React\Async\await;
 
 /**
  * Store device connection state message consumer
@@ -48,46 +55,54 @@ final class StoreDeviceConnectionState implements Queue\Consumer
 		private readonly DevicesModels\Configuration\Channels\Repository $channelsConfigurationRepository,
 		private readonly DevicesModels\Configuration\Channels\Properties\Repository $channelsPropertiesConfigurationRepository,
 		private readonly DevicesUtilities\DeviceConnection $deviceConnectionManager,
-		private readonly DevicesUtilities\DevicePropertiesStates $devicePropertiesStatesManager,
-		private readonly DevicesUtilities\ChannelPropertiesStates $channelPropertiesStatesManager,
+		private readonly DevicesModels\States\Async\DevicePropertiesManager $devicePropertiesStatesManager,
+		private readonly DevicesModels\States\Async\ChannelPropertiesManager $channelPropertiesStatesManager,
 	)
 	{
 	}
 
 	/**
+	 * @throws ApplicationExceptions\InvalidState
+	 * @throws ApplicationExceptions\Runtime
 	 * @throws DBAL\Exception
 	 * @throws DevicesExceptions\InvalidArgument
 	 * @throws DevicesExceptions\InvalidState
-	 * @throws DevicesExceptions\Runtime
 	 * @throws MetadataExceptions\InvalidArgument
 	 * @throws MetadataExceptions\InvalidState
+	 * @throws MetadataExceptions\Mapping
 	 * @throws MetadataExceptions\MalformedInput
+	 * @throws ToolsExceptions\InvalidArgument
+	 * @throws TypeError
+	 * @throws ValueError
 	 */
-	public function consume(Entities\Messages\Entity $entity): bool
+	public function consume(Queue\Messages\Message $message): bool
 	{
-		if (!$entity instanceof Entities\Messages\StoreDeviceConnectionState) {
+		if (!$message instanceof Queue\Messages\StoreDeviceConnectionState) {
 			return false;
 		}
 
-		$findDeviceQuery = new DevicesQueries\Configuration\FindDevices();
-		$findDeviceQuery->byConnectorId($entity->getConnector());
-		$findDeviceQuery->byIdentifier($entity->getIdentifier());
+		$findDeviceQuery = new Queries\Configuration\FindDevices();
+		$findDeviceQuery->byConnectorId($message->getConnector());
+		$findDeviceQuery->byIdentifier($message->getIdentifier());
 
-		$device = $this->devicesConfigurationRepository->findOneBy($findDeviceQuery);
+		$device = $this->devicesConfigurationRepository->findOneBy(
+			$findDeviceQuery,
+			Documents\Devices\Device::class,
+		);
 
 		if ($device === null) {
 			$this->logger->error(
 				'Device could not be loaded',
 				[
-					'source' => MetadataTypes\ConnectorSource::SOURCE_CONNECTOR_NS_PANEL,
+					'source' => MetadataTypes\Sources\Connector::NS_PANEL->value,
 					'type' => 'store-device-connection-state-message-consumer',
 					'connector' => [
-						'id' => $entity->getConnector()->toString(),
+						'id' => $message->getConnector()->toString(),
 					],
 					'device' => [
-						'identifier' => $entity->getIdentifier(),
+						'identifier' => $message->getIdentifier(),
 					],
-					'data' => $entity->toArray(),
+					'data' => $message->toArray(),
 				],
 			);
 
@@ -96,35 +111,42 @@ final class StoreDeviceConnectionState implements Queue\Consumer
 
 		// Check device state...
 		if (
-			!$this->deviceConnectionManager->getState($device)->equals($entity->getState())
+			$this->deviceConnectionManager->getState($device) !== $message->getState()
 		) {
 			// ... and if it is not ready, set it to ready
 			$this->deviceConnectionManager->setState(
 				$device,
-				$entity->getState(),
+				$message->getState(),
 			);
 
 			if (
-				$entity->getState()->equalsValue(MetadataTypes\ConnectionState::STATE_DISCONNECTED)
-				|| $entity->getState()->equalsValue(MetadataTypes\ConnectionState::STATE_ALERT)
-				|| $entity->getState()->equalsValue(MetadataTypes\ConnectionState::STATE_UNKNOWN)
+				$message->getState() === DevicesTypes\ConnectionState::DISCONNECTED
+				|| $message->getState() === DevicesTypes\ConnectionState::ALERT
+				|| $message->getState() === DevicesTypes\ConnectionState::UNKNOWN
 			) {
 				$findDevicePropertiesQuery = new DevicesQueries\Configuration\FindDeviceDynamicProperties();
 				$findDevicePropertiesQuery->forDevice($device);
 
 				$properties = $this->devicesPropertiesConfigurationRepository->findAllBy(
 					$findDevicePropertiesQuery,
-					MetadataDocuments\DevicesModule\DeviceDynamicProperty::class,
+					DevicesDocuments\Devices\Properties\Dynamic::class,
 				);
 
 				foreach ($properties as $property) {
-					$this->devicePropertiesStatesManager->setValidState($property, false);
+					await($this->devicePropertiesStatesManager->setValidState(
+						$property,
+						false,
+						MetadataTypes\Sources\Connector::NS_PANEL,
+					));
 				}
 
-				$findChannelsQuery = new DevicesQueries\Configuration\FindChannels();
+				$findChannelsQuery = new Queries\Configuration\FindChannels();
 				$findChannelsQuery->forDevice($device);
 
-				$channels = $this->channelsConfigurationRepository->findAllBy($findChannelsQuery);
+				$channels = $this->channelsConfigurationRepository->findAllBy(
+					$findChannelsQuery,
+					Documents\Channels\Channel::class,
+				);
 
 				foreach ($channels as $channel) {
 					$findChannelPropertiesQuery = new DevicesQueries\Configuration\FindChannelDynamicProperties();
@@ -132,31 +154,37 @@ final class StoreDeviceConnectionState implements Queue\Consumer
 
 					$properties = $this->channelsPropertiesConfigurationRepository->findAllBy(
 						$findChannelPropertiesQuery,
-						MetadataDocuments\DevicesModule\ChannelDynamicProperty::class,
+						DevicesDocuments\Channels\Properties\Dynamic::class,
 					);
 
 					foreach ($properties as $property) {
-						$this->channelPropertiesStatesManager->setValidState($property, false);
+						await($this->channelPropertiesStatesManager->setValidState(
+							$property,
+							false,
+							MetadataTypes\Sources\Connector::NS_PANEL,
+						));
 					}
 				}
 			}
 
-			if ($device->getType() === Entities\Devices\Gateway::TYPE) {
+			if ($device instanceof Documents\Devices\Gateway) {
 				if (
-					$entity->getState()->equalsValue(MetadataTypes\ConnectionState::STATE_DISCONNECTED)
-					|| $entity->getState()->equalsValue(MetadataTypes\ConnectionState::STATE_ALERT)
-					|| $entity->getState()->equalsValue(MetadataTypes\ConnectionState::STATE_UNKNOWN)
+					$message->getState() === DevicesTypes\ConnectionState::DISCONNECTED
+					|| $message->getState() === DevicesTypes\ConnectionState::ALERT
+					|| $message->getState() === DevicesTypes\ConnectionState::UNKNOWN
 				) {
-					$findChildrenDevicesQuery = new DevicesQueries\Configuration\FindDevices();
+					$findChildrenDevicesQuery = new Queries\Configuration\FindSubDevices();
 					$findChildrenDevicesQuery->forParent($device);
-					$findChildrenDevicesQuery->byType(Entities\Devices\SubDevice::TYPE);
 
-					$children = $this->devicesConfigurationRepository->findAllBy($findChildrenDevicesQuery);
+					$children = $this->devicesConfigurationRepository->findAllBy(
+						$findChildrenDevicesQuery,
+						Documents\Devices\SubDevice::class,
+					);
 
 					foreach ($children as $child) {
 						$this->deviceConnectionManager->setState(
 							$child,
-							$entity->getState(),
+							$message->getState(),
 						);
 
 						$findDevicePropertiesQuery = new DevicesQueries\Configuration\FindDeviceDynamicProperties();
@@ -164,17 +192,24 @@ final class StoreDeviceConnectionState implements Queue\Consumer
 
 						$properties = $this->devicesPropertiesConfigurationRepository->findAllBy(
 							$findDevicePropertiesQuery,
-							MetadataDocuments\DevicesModule\DeviceDynamicProperty::class,
+							DevicesDocuments\Devices\Properties\Dynamic::class,
 						);
 
 						foreach ($properties as $property) {
-							$this->devicePropertiesStatesManager->setValidState($property, false);
+							await($this->devicePropertiesStatesManager->setValidState(
+								$property,
+								false,
+								MetadataTypes\Sources\Connector::NS_PANEL,
+							));
 						}
 
-						$findChannelsQuery = new DevicesQueries\Configuration\FindChannels();
+						$findChannelsQuery = new Queries\Configuration\FindChannels();
 						$findChannelsQuery->forDevice($child);
 
-						$channels = $this->channelsConfigurationRepository->findAllBy($findChannelsQuery);
+						$channels = $this->channelsConfigurationRepository->findAllBy(
+							$findChannelsQuery,
+							Documents\Channels\Channel::class,
+						);
 
 						foreach ($channels as $channel) {
 							$findChannelPropertiesQuery = new DevicesQueries\Configuration\FindChannelDynamicProperties();
@@ -182,27 +217,33 @@ final class StoreDeviceConnectionState implements Queue\Consumer
 
 							$properties = $this->channelsPropertiesConfigurationRepository->findAllBy(
 								$findChannelPropertiesQuery,
-								MetadataDocuments\DevicesModule\ChannelDynamicProperty::class,
+								DevicesDocuments\Channels\Properties\Dynamic::class,
 							);
 
 							foreach ($properties as $property) {
-								$this->channelPropertiesStatesManager->setValidState($property, false);
+								await($this->channelPropertiesStatesManager->setValidState(
+									$property,
+									false,
+									MetadataTypes\Sources\Connector::NS_PANEL,
+								));
 							}
 						}
 					}
 				}
 
-				if ($entity->getState()->equalsValue(MetadataTypes\ConnectionState::STATE_ALERT)) {
-					$findChildrenDevicesQuery = new DevicesQueries\Configuration\FindDevices();
+				if ($message->getState() === DevicesTypes\ConnectionState::ALERT) {
+					$findChildrenDevicesQuery = new Queries\Configuration\FindThirdPartyDevices();
 					$findChildrenDevicesQuery->forParent($device);
-					$findChildrenDevicesQuery->byType(Entities\Devices\ThirdPartyDevice::TYPE);
 
-					$children = $this->devicesConfigurationRepository->findAllBy($findChildrenDevicesQuery);
+					$children = $this->devicesConfigurationRepository->findAllBy(
+						$findChildrenDevicesQuery,
+						Documents\Devices\ThirdPartyDevice::class,
+					);
 
 					foreach ($children as $child) {
 						$this->deviceConnectionManager->setState(
 							$child,
-							$entity->getState(),
+							$message->getState(),
 						);
 					}
 				}
@@ -212,15 +253,15 @@ final class StoreDeviceConnectionState implements Queue\Consumer
 		$this->logger->debug(
 			'Consumed device connection state message',
 			[
-				'source' => MetadataTypes\ConnectorSource::SOURCE_CONNECTOR_NS_PANEL,
+				'source' => MetadataTypes\Sources\Connector::NS_PANEL->value,
 				'type' => 'store-device-connection-state-message-consumer',
 				'connector' => [
-					'id' => $entity->getConnector()->toString(),
+					'id' => $message->getConnector()->toString(),
 				],
 				'device' => [
-					'identifier' => $entity->getIdentifier(),
+					'id' => $device->getId()->toString(),
 				],
-				'data' => $entity->toArray(),
+				'data' => $message->toArray(),
 			],
 		);
 
