@@ -15,28 +15,27 @@
 
 namespace FastyBird\Connector\NsPanel\Queue\Consumers;
 
-use DateTimeInterface;
 use Doctrine\DBAL;
 use FastyBird\Connector\NsPanel;
-use FastyBird\Connector\NsPanel\Entities;
+use FastyBird\Connector\NsPanel\Documents;
 use FastyBird\Connector\NsPanel\Helpers;
+use FastyBird\Connector\NsPanel\Models;
+use FastyBird\Connector\NsPanel\Queries;
 use FastyBird\Connector\NsPanel\Queue;
-use FastyBird\Library\Exchange\Documents as ExchangeEntities;
-use FastyBird\Library\Exchange\Exceptions as ExchangeExceptions;
-use FastyBird\Library\Exchange\Publisher as ExchangePublisher;
-use FastyBird\Library\Metadata\Documents as MetadataDocuments;
-use FastyBird\Library\Metadata\Exceptions as MetadataExceptions;
+use FastyBird\Library\Application\Exceptions as ApplicationExceptions;
+use FastyBird\Library\Application\Helpers as ApplicationHelpers;
 use FastyBird\Library\Metadata\Types as MetadataTypes;
 use FastyBird\Library\Metadata\Utilities as MetadataUtilities;
+use FastyBird\Module\Devices\Documents as DevicesDocuments;
 use FastyBird\Module\Devices\Entities as DevicesEntities;
 use FastyBird\Module\Devices\Exceptions as DevicesExceptions;
 use FastyBird\Module\Devices\Models as DevicesModels;
 use FastyBird\Module\Devices\Queries as DevicesQueries;
 use FastyBird\Module\Devices\States as DevicesStates;
-use FastyBird\Module\Devices\Utilities as DevicesUtilities;
 use Nette;
 use Nette\Utils;
 use function assert;
+use function React\Async\await;
 
 /**
  * Store device state message consumer
@@ -52,82 +51,77 @@ final class StoreDeviceState implements Queue\Consumer
 	use Nette\SmartObject;
 
 	public function __construct(
-		private readonly bool $useExchange,
 		private readonly NsPanel\Logger $logger,
+		private readonly Models\StateRepository $stateRepository,
 		private readonly DevicesModels\Entities\Channels\Properties\PropertiesRepository $channelsPropertiesRepository,
 		private readonly DevicesModels\Entities\Channels\Properties\PropertiesManager $channelsPropertiesManager,
 		private readonly DevicesModels\Configuration\Devices\Repository $devicesConfigurationRepository,
 		private readonly DevicesModels\Configuration\Channels\Repository $channelsConfigurationRepository,
 		private readonly DevicesModels\Configuration\Channels\Properties\Repository $channelsPropertiesConfigurationRepository,
-		private readonly DevicesUtilities\ChannelPropertiesStates $channelPropertiesStatesManager,
-		private readonly DevicesUtilities\Database $databaseHelper,
-		private readonly ExchangeEntities\DocumentFactory $entityFactory,
-		private readonly ExchangePublisher\Publisher $publisher,
+		private readonly DevicesModels\States\Async\ChannelPropertiesManager $channelPropertiesStatesManager,
+		private readonly ApplicationHelpers\Database $databaseHelper,
 	)
 	{
 	}
 
 	/**
+	 * @throws ApplicationExceptions\InvalidState
+	 * @throws ApplicationExceptions\Runtime
 	 * @throws DBAL\Exception
-	 * @throws DevicesExceptions\InvalidArgument
 	 * @throws DevicesExceptions\InvalidState
-	 * @throws DevicesExceptions\Runtime
-	 * @throws ExchangeExceptions\InvalidArgument
-	 * @throws ExchangeExceptions\InvalidState
-	 * @throws MetadataExceptions\InvalidArgument
-	 * @throws MetadataExceptions\InvalidState
-	 * @throws MetadataExceptions\MalformedInput
-	 * @throws Utils\JsonException
 	 */
-	public function consume(Entities\Messages\Entity $entity): bool
+	public function consume(Queue\Messages\Message $message): bool
 	{
-		if (!$entity instanceof Entities\Messages\StoreDeviceState) {
+		if (!$message instanceof Queue\Messages\StoreDeviceState) {
 			return false;
 		}
 
-		$findDeviceQuery = new DevicesQueries\Configuration\FindDevices();
-		$findDeviceQuery->byConnectorId($entity->getConnector());
-		$findDeviceQuery->byIdentifier($entity->getIdentifier());
+		$findDeviceQuery = new Queries\Configuration\FindDevices();
+		$findDeviceQuery->byConnectorId($message->getConnector());
+		$findDeviceQuery->byIdentifier($message->getIdentifier());
 
-		$device = $this->devicesConfigurationRepository->findOneBy($findDeviceQuery);
+		$device = $this->devicesConfigurationRepository->findOneBy(
+			$findDeviceQuery,
+			Documents\Devices\Device::class,
+		);
 
 		if ($device === null) {
 			$this->logger->error(
 				'Device could not be loaded',
 				[
-					'source' => MetadataTypes\ConnectorSource::SOURCE_CONNECTOR_NS_PANEL,
+					'source' => MetadataTypes\Sources\Connector::NS_PANEL->value,
 					'type' => 'store-device-state-message-consumer',
 					'connector' => [
-						'id' => $entity->getConnector()->toString(),
+						'id' => $message->getConnector()->toString(),
 					],
 					'device' => [
-						'identifier' => $entity->getIdentifier(),
+						'identifier' => $message->getIdentifier(),
 					],
-					'data' => $entity->toArray(),
+					'data' => $message->toArray(),
 				],
 			);
 
 			return true;
 		}
 
-		if ($device->getType() === Entities\Devices\ThirdPartyDevice::TYPE) {
-			$this->processThirdPartyDevice($device, $entity->getState());
-		} elseif ($device->getType() === Entities\Devices\SubDevice::TYPE) {
-			$this->processSubDevice($device, $entity->getState());
+		if ($device instanceof Documents\Devices\ThirdPartyDevice) {
+			$this->processThirdPartyDevice($device, $message->getState());
+		} elseif ($device instanceof Documents\Devices\SubDevice) {
+			$this->processSubDevice($device, $message->getState());
 		}
 
 		$this->logger->debug(
 			'Consumed store device state message',
 			[
-				'source' => MetadataTypes\ConnectorSource::SOURCE_CONNECTOR_NS_PANEL,
+				'source' => MetadataTypes\Sources\Connector::NS_PANEL->value,
 				'type' => 'store-device-state-message-consumer',
 				'connector' => [
-					'id' => $entity->getConnector()->toString(),
+					'id' => $message->getConnector()->toString(),
 				],
 				'device' => [
-					'identifier' => $entity->getIdentifier(),
+					'id' => $device->getId()->toString(),
 				],
-				'data' => $entity->toArray(),
+				'data' => $message->toArray(),
 			],
 		);
 
@@ -135,27 +129,23 @@ final class StoreDeviceState implements Queue\Consumer
 	}
 
 	/**
-	 * @param array<Entities\Messages\CapabilityState> $state
+	 * @param array<Queue\Messages\CapabilityState> $state
 	 *
-	 * @throws DevicesExceptions\InvalidArgument
 	 * @throws DevicesExceptions\InvalidState
-	 * @throws MetadataExceptions\InvalidArgument
-	 * @throws MetadataExceptions\InvalidState
-	 * @throws MetadataExceptions\MalformedInput
 	 */
-	private function processSubDevice(
-		MetadataDocuments\DevicesModule\Device $device,
-		array $state,
-	): void
+	private function processSubDevice(Documents\Devices\Device $device, array $state): void
 	{
 		foreach ($state as $item) {
-			$findChannelQuery = new DevicesQueries\Configuration\FindChannels();
+			$findChannelQuery = new Queries\Configuration\FindChannels();
 			$findChannelQuery->forDevice($device);
 			$findChannelQuery->byIdentifier(
 				Helpers\Name::convertCapabilityToChannel($item->getCapability(), $item->getIdentifier()),
 			);
 
-			$channel = $this->channelsConfigurationRepository->findOneBy($findChannelQuery);
+			$channel = $this->channelsConfigurationRepository->findOneBy(
+				$findChannelQuery,
+				Documents\Channels\Channel::class,
+			);
 
 			if ($channel === null) {
 				continue;
@@ -167,54 +157,49 @@ final class StoreDeviceState implements Queue\Consumer
 
 			$property = $this->channelsPropertiesConfigurationRepository->findOneBy(
 				$findChannelPropertiesQuery,
-				MetadataDocuments\DevicesModule\ChannelDynamicProperty::class,
+				DevicesDocuments\Channels\Properties\Dynamic::class,
 			);
 
 			if ($property === null) {
 				continue;
 			}
 
-			$this->channelPropertiesStatesManager->writeValue(
+			await($this->channelPropertiesStatesManager->set(
 				$property,
 				Utils\ArrayHash::from([
-					DevicesStates\Property::ACTUAL_VALUE_FIELD => MetadataUtilities\ValueHelper::transformValueFromDevice(
-						$property->getDataType(),
-						$property->getFormat(),
-						MetadataUtilities\ValueHelper::flattenValue($item->getValue()),
-					),
-					DevicesStates\Property::VALID_FIELD => true,
+					DevicesStates\Property::ACTUAL_VALUE_FIELD => $item->getValue(),
 				]),
-			);
+				MetadataTypes\Sources\Connector::NS_PANEL,
+			));
+
+			$this->stateRepository->set($property->getId(), MetadataUtilities\Value::flattenValue($item->getValue()));
 		}
 	}
 
 	/**
-	 * @param array<Entities\Messages\CapabilityState> $state
+	 * @param array<Queue\Messages\CapabilityState> $state
 	 *
+	 * @throws ApplicationExceptions\InvalidState
+	 * @throws ApplicationExceptions\Runtime
 	 * @throws DBAL\Exception
-	 * @throws DevicesExceptions\InvalidArgument
 	 * @throws DevicesExceptions\InvalidState
-	 * @throws DevicesExceptions\Runtime
-	 * @throws ExchangeExceptions\InvalidArgument
-	 * @throws ExchangeExceptions\InvalidState
-	 * @throws MetadataExceptions\InvalidArgument
-	 * @throws MetadataExceptions\InvalidState
-	 * @throws MetadataExceptions\MalformedInput
-	 * @throws Utils\JsonException
 	 */
 	private function processThirdPartyDevice(
-		MetadataDocuments\DevicesModule\Device $device,
+		Documents\Devices\Device $device,
 		array $state,
 	): void
 	{
 		foreach ($state as $item) {
-			$findChannelQuery = new DevicesQueries\Configuration\FindChannels();
+			$findChannelQuery = new Queries\Configuration\FindChannels();
 			$findChannelQuery->forDevice($device);
 			$findChannelQuery->byIdentifier(
 				Helpers\Name::convertCapabilityToChannel($item->getCapability(), $item->getIdentifier()),
 			);
 
-			$channel = $this->channelsConfigurationRepository->findOneBy($findChannelQuery);
+			$channel = $this->channelsConfigurationRepository->findOneBy(
+				$findChannelQuery,
+				Documents\Channels\Channel::class,
+			);
 
 			if ($channel === null) {
 				continue;
@@ -230,36 +215,25 @@ final class StoreDeviceState implements Queue\Consumer
 				continue;
 			}
 
-			$value = MetadataUtilities\ValueHelper::transformValueFromDevice(
-				$property->getDataType(),
-				$property->getFormat(),
-				MetadataUtilities\ValueHelper::flattenValue($item->getValue()),
+			$this->writeThirdPartyProperty(
+				$property,
+				MetadataUtilities\Value::flattenValue($item->getValue()),
 			);
-
-			$this->writeThirdPartyProperty($device, $channel, $property, $value);
 		}
 	}
 
 	/**
+	 * @throws ApplicationExceptions\InvalidState
+	 * @throws ApplicationExceptions\Runtime
 	 * @throws DBAL\Exception
-	 * @throws DevicesExceptions\InvalidArgument
 	 * @throws DevicesExceptions\InvalidState
-	 * @throws DevicesExceptions\Runtime
-	 * @throws ExchangeExceptions\InvalidArgument
-	 * @throws ExchangeExceptions\InvalidState
-	 * @throws MetadataExceptions\InvalidArgument
-	 * @throws MetadataExceptions\InvalidState
-	 * @throws MetadataExceptions\MalformedInput
-	 * @throws Utils\JsonException
 	 */
 	private function writeThirdPartyProperty(
-		MetadataDocuments\DevicesModule\Device $device,
-		MetadataDocuments\DevicesModule\Channel $channel,
-		MetadataDocuments\DevicesModule\ChannelProperty $property,
-		float|int|string|bool|MetadataTypes\ButtonPayload|MetadataTypes\CoverPayload|MetadataTypes\SwitchPayload|DateTimeInterface|null $value,
+		DevicesDocuments\Channels\Properties\Property $property,
+		float|int|string|bool|null $value,
 	): void
 	{
-		if ($property instanceof MetadataDocuments\DevicesModule\ChannelVariableProperty) {
+		if ($property instanceof DevicesDocuments\Channels\Properties\Variable) {
 			$this->databaseHelper->transaction(
 				function () use ($property, $value): void {
 					$property = $this->channelsPropertiesRepository->find(
@@ -277,50 +251,27 @@ final class StoreDeviceState implements Queue\Consumer
 				},
 			);
 
-		} elseif ($property instanceof MetadataDocuments\DevicesModule\ChannelDynamicProperty) {
-			$this->channelPropertiesStatesManager->writeValue(
+		} elseif ($property instanceof DevicesDocuments\Channels\Properties\Dynamic) {
+			await($this->channelPropertiesStatesManager->set(
 				$property,
 				Utils\ArrayHash::from([
-					DevicesStates\Property::ACTUAL_VALUE_FIELD => MetadataUtilities\ValueHelper::transformValueFromDevice(
-						$property->getDataType(),
-						$property->getFormat(),
-						MetadataUtilities\ValueHelper::flattenValue($value),
-					),
-					DevicesStates\Property::VALID_FIELD => true,
+					DevicesStates\Property::ACTUAL_VALUE_FIELD => $value,
 				]),
-			);
+				MetadataTypes\Sources\Connector::NS_PANEL,
+			));
 
-		} elseif ($property instanceof MetadataDocuments\DevicesModule\ChannelMappedProperty) {
-			if ($this->useExchange) {
-				$this->publisher->publish(
-					MetadataTypes\ConnectorSource::get(
-						MetadataTypes\ConnectorSource::SOURCE_CONNECTOR_NS_PANEL,
-					),
-					MetadataTypes\RoutingKey::get(
-						MetadataTypes\RoutingKey::CHANNEL_PROPERTY_ACTION,
-					),
-					$this->entityFactory->create(
-						Utils\Json::encode([
-							'action' => MetadataTypes\PropertyAction::ACTION_SET,
-							'device' => $device->getId()->toString(),
-							'channel' => $channel->getId()->toString(),
-							'property' => $property->getId()->toString(),
-							'expected_value' => MetadataUtilities\ValueHelper::flattenValue($value),
-						]),
-						MetadataTypes\RoutingKey::get(
-							MetadataTypes\RoutingKey::CHANNEL_PROPERTY_ACTION,
-						),
-					),
-				);
-			} else {
-				$this->channelPropertiesStatesManager->writeValue(
-					$property,
-					Utils\ArrayHash::from([
-						DevicesStates\Property::EXPECTED_VALUE_FIELD => $value,
-						DevicesStates\Property::PENDING_FIELD => true,
-					]),
-				);
-			}
+			$this->stateRepository->set($property->getId(), $value);
+
+		} elseif ($property instanceof DevicesDocuments\Channels\Properties\Mapped) {
+			await($this->channelPropertiesStatesManager->write(
+				$property,
+				Utils\ArrayHash::from([
+					DevicesStates\Property::EXPECTED_VALUE_FIELD => $value,
+				]),
+				MetadataTypes\Sources\Connector::NS_PANEL,
+			));
+
+			$this->stateRepository->set($property->getId(), $value);
 		}
 	}
 
