@@ -18,12 +18,14 @@ namespace FastyBird\Connector\NsPanel\Queue\Consumers;
 use Doctrine\DBAL;
 use FastyBird\Connector\NsPanel;
 use FastyBird\Connector\NsPanel\Entities;
+use FastyBird\Connector\NsPanel\Exceptions;
 use FastyBird\Connector\NsPanel\Helpers;
 use FastyBird\Connector\NsPanel\Queries;
 use FastyBird\Connector\NsPanel\Queue;
 use FastyBird\Connector\NsPanel\Types;
+use FastyBird\Library\Application\Exceptions as ApplicationExceptions;
+use FastyBird\Library\Application\Helpers as ApplicationHelpers;
 use FastyBird\Library\Metadata\Types as MetadataTypes;
-use FastyBird\Module\Devices\Exceptions as DevicesExceptions;
 use FastyBird\Module\Devices\Models as DevicesModels;
 use FastyBird\Module\Devices\Utilities as DevicesUtilities;
 use Nette;
@@ -50,7 +52,7 @@ final class StoreSubDevice implements Queue\Consumer
 		protected readonly DevicesModels\Entities\Devices\DevicesRepository $devicesRepository,
 		protected readonly DevicesModels\Entities\Devices\Properties\PropertiesRepository $devicesPropertiesRepository,
 		protected readonly DevicesModels\Entities\Devices\Properties\PropertiesManager $devicesPropertiesManager,
-		protected readonly DevicesUtilities\Database $databaseHelper,
+		protected readonly ApplicationHelpers\Database $databaseHelper,
 		private readonly DevicesModels\Entities\Connectors\ConnectorsRepository $connectorsRepository,
 		private readonly DevicesModels\Entities\Devices\DevicesManager $devicesManager,
 		private readonly DevicesModels\Entities\Channels\ChannelsRepository $channelsRepository,
@@ -60,34 +62,35 @@ final class StoreSubDevice implements Queue\Consumer
 	}
 
 	/**
+	 * @throws ApplicationExceptions\InvalidState
+	 * @throws ApplicationExceptions\Runtime
+	 * @throws Exceptions\InvalidArgument
 	 * @throws DBAL\Exception
-	 * @throws DevicesExceptions\InvalidState
-	 * @throws DevicesExceptions\Runtime
 	 */
-	public function consume(Entities\Messages\Entity $entity): bool
+	public function consume(Queue\Messages\Message $message): bool
 	{
-		if (!$entity instanceof Entities\Messages\StoreSubDevice) {
+		if (!$message instanceof Queue\Messages\StoreSubDevice) {
 			return false;
 		}
 
-		$parent = $this->devicesRepository->find($entity->getGateway(), Entities\Devices\Gateway::class);
+		$parent = $this->devicesRepository->find($message->getGateway(), Entities\Devices\Gateway::class);
 
 		if ($parent === null) {
 			$this->logger->error(
 				'Device could not be loaded',
 				[
-					'source' => MetadataTypes\ConnectorSource::SOURCE_CONNECTOR_NS_PANEL,
+					'source' => MetadataTypes\Sources\Connector::NS_PANEL->value,
 					'type' => 'store-sub-device-message-consumer',
 					'connector' => [
-						'id' => $entity->getConnector()->toString(),
+						'id' => $message->getConnector()->toString(),
 					],
 					'gateway' => [
-						'id' => $entity->getGateway()->toString(),
+						'id' => $message->getGateway()->toString(),
 					],
 					'device' => [
-						'identifier' => $entity->getIdentifier(),
+						'identifier' => $message->getIdentifier(),
 					],
-					'data' => $entity->toArray(),
+					'data' => $message->toArray(),
 				],
 			);
 
@@ -95,34 +98,34 @@ final class StoreSubDevice implements Queue\Consumer
 		}
 
 		$findDeviceQuery = new Queries\Entities\FindSubDevices();
-		$findDeviceQuery->byConnectorId($entity->getConnector());
+		$findDeviceQuery->byConnectorId($message->getConnector());
 		$findDeviceQuery->forParent($parent);
-		$findDeviceQuery->byIdentifier($entity->getIdentifier());
+		$findDeviceQuery->byIdentifier($message->getIdentifier());
 
 		$device = $this->devicesRepository->findOneBy($findDeviceQuery, Entities\Devices\SubDevice::class);
 
 		if ($device === null) {
 			$connector = $this->connectorsRepository->find(
-				$entity->getConnector(),
-				Entities\NsPanelConnector::class,
+				$message->getConnector(),
+				Entities\Connectors\Connector::class,
 			);
 
 			if ($connector === null) {
 				$this->logger->error(
 					'Connector could not be loaded',
 					[
-						'source' => MetadataTypes\ConnectorSource::SOURCE_CONNECTOR_NS_PANEL,
+						'source' => MetadataTypes\Sources\Connector::NS_PANEL->value,
 						'type' => 'store-sub-device-message-consumer',
 						'connector' => [
-							'id' => $entity->getConnector()->toString(),
+							'id' => $message->getConnector()->toString(),
 						],
 						'gateway' => [
-							'id' => $entity->getGateway()->toString(),
+							'id' => $message->getGateway()->toString(),
 						],
 						'device' => [
-							'identifier' => $entity->getIdentifier(),
+							'identifier' => $message->getIdentifier(),
 						],
-						'data' => $entity->toArray(),
+						'data' => $message->toArray(),
 					],
 				);
 
@@ -130,13 +133,13 @@ final class StoreSubDevice implements Queue\Consumer
 			}
 
 			$device = $this->databaseHelper->transaction(
-				function () use ($entity, $connector, $parent): Entities\Devices\SubDevice {
+				function () use ($message, $connector, $parent): Entities\Devices\SubDevice {
 					$device = $this->devicesManager->create(Utils\ArrayHash::from([
 						'entity' => Entities\Devices\SubDevice::class,
 						'connector' => $connector,
 						'parent' => $parent,
-						'identifier' => $entity->getIdentifier(),
-						'name' => $entity->getName(),
+						'identifier' => $message->getIdentifier(),
+						'name' => $message->getName(),
 					]));
 					assert($device instanceof Entities\Devices\SubDevice);
 
@@ -147,18 +150,18 @@ final class StoreSubDevice implements Queue\Consumer
 			$this->logger->info(
 				'Sub-device was created',
 				[
-					'source' => MetadataTypes\ConnectorSource::SOURCE_CONNECTOR_NS_PANEL,
+					'source' => MetadataTypes\Sources\Connector::NS_PANEL->value,
 					'type' => 'store-sub-device-message-consumer',
 					'connector' => [
-						'id' => $entity->getConnector()->toString(),
+						'id' => $connector->getId()->toString(),
 					],
 					'gateway' => [
-						'id' => $entity->getGateway()->toString(),
+						'id' => $message->getGateway()->toString(),
 					],
 					'device' => [
 						'id' => $device->getId()->toString(),
-						'identifier' => $entity->getIdentifier(),
-						'protocol' => $entity->getProtocol(),
+						'identifier' => $message->getIdentifier(),
+						'protocol' => $message->getProtocol(),
 					],
 				],
 			);
@@ -166,46 +169,46 @@ final class StoreSubDevice implements Queue\Consumer
 
 		$this->setDeviceProperty(
 			$device->getId(),
-			$entity->getManufacturer(),
-			MetadataTypes\DataType::get(MetadataTypes\DataType::DATA_TYPE_STRING),
+			$message->getManufacturer(),
+			MetadataTypes\DataType::STRING,
 			Types\DevicePropertyIdentifier::MANUFACTURER,
-			DevicesUtilities\Name::createName(Types\DevicePropertyIdentifier::MANUFACTURER),
+			DevicesUtilities\Name::createName(Types\DevicePropertyIdentifier::MANUFACTURER->value),
 		);
 		$this->setDeviceProperty(
 			$device->getId(),
-			$entity->getModel(),
-			MetadataTypes\DataType::get(MetadataTypes\DataType::DATA_TYPE_STRING),
+			$message->getModel(),
+			MetadataTypes\DataType::STRING,
 			Types\DevicePropertyIdentifier::MODEL,
-			DevicesUtilities\Name::createName(Types\DevicePropertyIdentifier::MODEL),
+			DevicesUtilities\Name::createName(Types\DevicePropertyIdentifier::MODEL->value),
 		);
 		$this->setDeviceProperty(
 			$device->getId(),
-			$entity->getFirmwareVersion(),
-			MetadataTypes\DataType::get(MetadataTypes\DataType::DATA_TYPE_STRING),
+			$message->getFirmwareVersion(),
+			MetadataTypes\DataType::STRING,
 			Types\DevicePropertyIdentifier::FIRMWARE_VERSION,
-			DevicesUtilities\Name::createName(Types\DevicePropertyIdentifier::FIRMWARE_VERSION),
+			DevicesUtilities\Name::createName(Types\DevicePropertyIdentifier::FIRMWARE_VERSION->value),
 		);
 		$this->setDeviceProperty(
 			$device->getId(),
-			$entity->getDisplayCategory()->getValue(),
-			MetadataTypes\DataType::get(MetadataTypes\DataType::DATA_TYPE_STRING),
+			$message->getDisplayCategory()->value,
+			MetadataTypes\DataType::STRING,
 			Types\DevicePropertyIdentifier::CATEGORY,
-			DevicesUtilities\Name::createName(Types\DevicePropertyIdentifier::CATEGORY),
+			DevicesUtilities\Name::createName(Types\DevicePropertyIdentifier::CATEGORY->value),
 		);
 		$this->setDeviceProperty(
 			$device->getId(),
-			$entity->getMacAddress(),
-			MetadataTypes\DataType::get(MetadataTypes\DataType::DATA_TYPE_STRING),
+			$message->getMacAddress(),
+			MetadataTypes\DataType::STRING,
 			Types\DevicePropertyIdentifier::MAC_ADDRESS,
-			DevicesUtilities\Name::createName(Types\DevicePropertyIdentifier::MAC_ADDRESS),
+			DevicesUtilities\Name::createName(Types\DevicePropertyIdentifier::MAC_ADDRESS->value),
 		);
 
-		foreach ($entity->getCapabilities() as $capability) {
-			$this->databaseHelper->transaction(function () use ($entity, $device, $capability): bool {
+		foreach ($message->getCapabilities() as $capability) {
+			$this->databaseHelper->transaction(function () use ($message, $device, $capability): bool {
 				$identifier = Helpers\Name::convertCapabilityToChannel($capability->getCapability());
 
 				if (
-					$capability->getCapability()->equalsValue(Types\Capability::TOGGLE)
+					$capability->getCapability() === Types\Capability::TOGGLE
 					&& $capability->getName() !== null
 				) {
 					$identifier .= '_' . $capability->getName();
@@ -215,11 +218,11 @@ final class StoreSubDevice implements Queue\Consumer
 				$findChannelQuery->byIdentifier($identifier);
 				$findChannelQuery->forDevice($device);
 
-				$channel = $this->channelsRepository->findOneBy($findChannelQuery, Entities\NsPanelChannel::class);
+				$channel = $this->channelsRepository->findOneBy($findChannelQuery, Entities\Channels\Channel::class);
 
 				if ($channel === null) {
 					$channel = $this->channelsManager->create(Utils\ArrayHash::from([
-						'entity' => Entities\NsPanelChannel::class,
+						'entity' => Entities\Channels\Channel::class,
 						'device' => $device,
 						'identifier' => $identifier,
 					]));
@@ -227,16 +230,16 @@ final class StoreSubDevice implements Queue\Consumer
 					$this->logger->debug(
 						'Device channel was created',
 						[
-							'source' => MetadataTypes\ConnectorSource::SOURCE_CONNECTOR_NS_PANEL,
+							'source' => MetadataTypes\Sources\Connector::NS_PANEL->value,
 							'type' => 'store-sub-device-message-consumer',
 							'connector' => [
-								'id' => $entity->getConnector()->toString(),
+								'id' => $message->getConnector()->toString(),
 							],
 							'gateway' => [
-								'id' => $entity->getGateway()->toString(),
+								'id' => $message->getGateway()->toString(),
 							],
 							'device' => [
-								'identifier' => $entity->getIdentifier(),
+								'id' => $device->getId()->toString(),
 							],
 							'channel' => [
 								'id' => $channel->getId()->toString(),
@@ -249,14 +252,14 @@ final class StoreSubDevice implements Queue\Consumer
 			});
 		}
 
-		foreach ($entity->getTags() as $tag => $value) {
-			if ($tag === Types\Capability::TOGGLE && is_array($value)) {
-				$this->databaseHelper->transaction(function () use ($entity, $device, $value): void {
+		foreach ($message->getTags() as $tag => $value) {
+			if ($tag === Types\Capability::TOGGLE->value && is_array($value)) {
+				$this->databaseHelper->transaction(function () use ($message, $device, $value): void {
 					foreach ($value as $key => $name) {
 						$findChannelQuery = new Queries\Entities\FindChannels();
 						$findChannelQuery->byIdentifier(
 							Helpers\Name::convertCapabilityToChannel(
-								Types\Capability::get(Types\Capability::TOGGLE),
+								Types\Capability::TOGGLE,
 								$key,
 							),
 						);
@@ -264,7 +267,7 @@ final class StoreSubDevice implements Queue\Consumer
 
 						$channel = $this->channelsRepository->findOneBy(
 							$findChannelQuery,
-							Entities\NsPanelChannel::class,
+							Entities\Channels\Channel::class,
 						);
 
 						if ($channel !== null) {
@@ -275,16 +278,16 @@ final class StoreSubDevice implements Queue\Consumer
 							$this->logger->debug(
 								'Toggle channel name was set',
 								[
-									'source' => MetadataTypes\ConnectorSource::SOURCE_CONNECTOR_NS_PANEL,
+									'source' => MetadataTypes\Sources\Connector::NS_PANEL->value,
 									'type' => 'store-sub-device-message-consumer',
 									'connector' => [
-										'id' => $entity->getConnector()->toString(),
+										'id' => $message->getConnector()->toString(),
 									],
 									'gateway' => [
-										'id' => $entity->getGateway()->toString(),
+										'id' => $message->getGateway()->toString(),
 									],
 									'device' => [
-										'identifier' => $entity->getIdentifier(),
+										'id' => $device->getId()->toString(),
 									],
 									'channel' => [
 										'id' => $channel->getId()->toString(),
@@ -300,18 +303,18 @@ final class StoreSubDevice implements Queue\Consumer
 		$this->logger->debug(
 			'Consumed store device message',
 			[
-				'source' => MetadataTypes\ConnectorSource::SOURCE_CONNECTOR_NS_PANEL,
+				'source' => MetadataTypes\Sources\Connector::NS_PANEL->value,
 				'type' => 'store-sub-device-message-consumer',
 				'connector' => [
-					'id' => $entity->getConnector()->toString(),
+					'id' => $message->getConnector()->toString(),
 				],
 				'gateway' => [
-					'id' => $entity->getGateway()->toString(),
+					'id' => $message->getGateway()->toString(),
 				],
 				'device' => [
-					'identifier' => $entity->getIdentifier(),
+					'id' => $device->getId()->toString(),
 				],
-				'data' => $entity->toArray(),
+				'data' => $message->toArray(),
 			],
 		);
 
