@@ -15,7 +15,7 @@
 
 namespace FastyBird\Connector\NsPanel\Servers;
 
-use FastyBird\Library\Metadata\Documents as MetadataDocuments;
+use FastyBird\Connector\NsPanel\Documents;
 
 /**
  * Connector base communication server factory
@@ -28,6 +28,6 @@ use FastyBird\Library\Metadata\Documents as MetadataDocuments;
 interface ServerFactory
 {
 
-	public function create(MetadataDocuments\DevicesModule\Connector $connector): Server;
+	public function create(Documents\Connectors\Connector $connector): Server;
 
 }
