@@ -19,10 +19,10 @@ use Doctrine\Common;
 use Doctrine\ORM;
 use Doctrine\Persistence;
 use FastyBird\Connector\NsPanel\Entities;
-use FastyBird\Connector\NsPanel\Types;
-use FastyBird\Module\Devices\Exceptions as DevicesExceptions;
+use FastyBird\Library\Application\Exceptions as ApplicationExceptions;
 use FastyBird\Module\Devices\Models as DevicesModels;
 use FastyBird\Module\Devices\Queries as DevicesQueries;
+use FastyBird\Module\Devices\Types as DevicesTypes;
 use Nette;
 use Nette\Utils;
 
@@ -56,36 +56,36 @@ final class Controls implements Common\EventSubscriber
 	/**
 	 * @param Persistence\Event\LifecycleEventArgs<ORM\EntityManagerInterface> $eventArgs
 	 *
-	 * @throws DevicesExceptions\InvalidState
+	 * @throws ApplicationExceptions\InvalidState
 	 */
 	public function postPersist(Persistence\Event\LifecycleEventArgs $eventArgs): void
 	{
 		$entity = $eventArgs->getObject();
 
 		// Check for valid entity
-		if ($entity instanceof Entities\NsPanelConnector) {
+		if ($entity instanceof Entities\Connectors\Connector) {
 			$findConnectorControlQuery = new DevicesQueries\Entities\FindConnectorControls();
 			$findConnectorControlQuery->forConnector($entity);
-			$findConnectorControlQuery->byName(Types\ConnectorControlName::DISCOVER);
+			$findConnectorControlQuery->byName(DevicesTypes\ControlName::DISCOVER->value);
 
 			$discoveryControl = $this->connectorsControlsRepository->findOneBy($findConnectorControlQuery);
 
 			if ($discoveryControl === null) {
 				$this->connectorsControlsManager->create(Utils\ArrayHash::from([
-					'name' => Types\ConnectorControlName::DISCOVER,
+					'name' => DevicesTypes\ControlName::DISCOVER->value,
 					'connector' => $entity,
 				]));
 			}
 
 			$findConnectorControlQuery = new DevicesQueries\Entities\FindConnectorControls();
 			$findConnectorControlQuery->forConnector($entity);
-			$findConnectorControlQuery->byName(Types\ConnectorControlName::REBOOT);
+			$findConnectorControlQuery->byName(DevicesTypes\ControlName::REBOOT->value);
 
 			$rebootControl = $this->connectorsControlsRepository->findOneBy($findConnectorControlQuery);
 
 			if ($rebootControl === null) {
 				$this->connectorsControlsManager->create(Utils\ArrayHash::from([
-					'name' => Types\ConnectorControlName::REBOOT,
+					'name' => DevicesTypes\ControlName::REBOOT->value,
 					'connector' => $entity,
 				]));
 			}
