@@ -15,15 +15,18 @@
 
 namespace FastyBird\Connector\NsPanel\Writers;
 
-use FastyBird\Connector\NsPanel\Entities;
+use FastyBird\Connector\NsPanel\Documents;
 use FastyBird\Connector\NsPanel\Exceptions;
-use FastyBird\Library\Metadata\Documents as MetadataDocuments;
+use FastyBird\Connector\NsPanel\Queries;
+use FastyBird\Connector\NsPanel\Queue;
 use FastyBird\Library\Metadata\Exceptions as MetadataExceptions;
-use FastyBird\Library\Metadata\Types as MetadataTypes;
 use FastyBird\Module\Devices\Events as DevicesEvents;
 use FastyBird\Module\Devices\Exceptions as DevicesExceptions;
 use FastyBird\Module\Devices\Queries as DevicesQueries;
+use FastyBird\Module\Devices\Types as DevicesTypes;
 use Symfony\Component\EventDispatcher;
+use TypeError;
+use ValueError;
 
 /**
  * Event based properties writer
@@ -58,28 +61,31 @@ class Event extends Periodic implements Writer, EventDispatcher\EventSubscriberI
 
 	/**
 	 * @throws DevicesExceptions\InvalidState
+	 * @throws Exceptions\InvalidArgument
 	 * @throws Exceptions\Runtime
 	 * @throws MetadataExceptions\InvalidArgument
 	 * @throws MetadataExceptions\InvalidState
+	 * @throws TypeError
+	 * @throws ValueError
 	 */
 	public function stateChanged(
 		DevicesEvents\ChannelPropertyStateEntityCreated|DevicesEvents\ChannelPropertyStateEntityUpdated $event,
 	): void
 	{
-		$property = $event->getProperty();
+		$findChannelQuery = new Queries\Configuration\FindChannels();
+		$findChannelQuery->byId($event->getProperty()->getChannel());
 
-		$state = $event->getState();
-
-		$findChannelQuery = new DevicesQueries\Configuration\FindChannels();
-		$findChannelQuery->byId($property->getChannel());
-
-		$channel = $this->channelsConfigurationRepository->findOneBy($findChannelQuery);
+		$channel = $this->channelsConfigurationRepository->findOneBy(
+			$findChannelQuery,
+			Documents\Channels\Channel::class,
+		);
 
 		if ($channel === null) {
 			return;
 		}
 
 		$findDeviceQuery = new DevicesQueries\Configuration\FindDevices();
+		$findDeviceQuery->forConnector($this->connector);
 		$findDeviceQuery->byId($channel->getDevice());
 
 		$device = $this->devicesConfigurationRepository->findOneBy($findDeviceQuery);
@@ -88,78 +94,49 @@ class Event extends Periodic implements Writer, EventDispatcher\EventSubscriberI
 			return;
 		}
 
-		if ($device->getType() === Entities\Devices\SubDevice::TYPE) {
-			if ($state->getExpectedValue() === null || $state->getPending() !== true) {
-				return;
-			}
-
-			$this->writeSubDeviceChannel($device, $channel);
-
-		} elseif ($device->getType() === Entities\Devices\ThirdPartyDevice::TYPE) {
-			if ($state->isValid() !== true) {
-				return;
-			}
-
-			$this->writeThirdPartyDeviceChannel($device, $channel);
-		}
-	}
-
-	/**
-	 * @throws Exceptions\Runtime
-	 */
-	public function writeSubDeviceChannel(
-		MetadataDocuments\DevicesModule\Device $device,
-		MetadataDocuments\DevicesModule\Channel $channel,
-	): void
-	{
-		$this->queue->append(
-			$this->entityHelper->create(
-				Entities\Messages\WriteSubDeviceState::class,
-				[
-					'connector' => $device->getConnector(),
-					'device' => $device->getId(),
-					'channel' => $channel->getId(),
-				],
-			),
-		);
-	}
-
-	/**
-	 * @throws DevicesExceptions\InvalidState
-	 * @throws Exceptions\Runtime
-	 * @throws MetadataExceptions\InvalidArgument
-	 * @throws MetadataExceptions\InvalidState
-	 */
-	public function writeThirdPartyDeviceChannel(
-		MetadataDocuments\DevicesModule\Device $device,
-		MetadataDocuments\DevicesModule\Channel $channel,
-	): void
-	{
-		if ($this->thirdPartyDeviceHelper->getGatewayIdentifier($device) === null) {
+		if ($device instanceof Documents\Devices\SubDevice) {
 			$this->queue->append(
-				$this->entityHelper->create(
-					Entities\Messages\StoreDeviceConnectionState::class,
+				$this->messageBuilder->create(
+					Queue\Messages\WriteSubDeviceState::class,
 					[
 						'connector' => $device->getConnector(),
-						'identifier' => $device->getIdentifier(),
-						'state' => MetadataTypes\ConnectionState::STATE_ALERT,
+						'device' => $device->getId(),
+						'channel' => $channel->getId(),
+						'property' => $event->getProperty()->getId(),
+						'state' => $event->getGet()->toArray(),
 					],
 				),
 			);
 
-			return;
-		}
+		} elseif ($device instanceof Documents\Devices\ThirdPartyDevice) {
+			if ($this->thirdPartyDeviceHelper->getGatewayIdentifier($device) === null) {
+				$this->queue->append(
+					$this->messageBuilder->create(
+						Queue\Messages\StoreDeviceConnectionState::class,
+						[
+							'connector' => $device->getConnector(),
+							'identifier' => $device->getIdentifier(),
+							'state' => DevicesTypes\ConnectionState::ALERT,
+						],
+					),
+				);
 
-		$this->queue->append(
-			$this->entityHelper->create(
-				Entities\Messages\WriteThirdPartyDeviceState::class,
-				[
-					'connector' => $device->getConnector(),
-					'device' => $device->getId(),
-					'channel' => $channel->getId(),
-				],
-			),
-		);
+				return;
+			}
+
+			$this->queue->append(
+				$this->messageBuilder->create(
+					Queue\Messages\WriteThirdPartyDeviceState::class,
+					[
+						'connector' => $device->getConnector(),
+						'device' => $device->getId(),
+						'channel' => $channel->getId(),
+						'property' => $event->getProperty()->getId(),
+						'state' => $event->getRead()->toArray(),
+					],
+				),
+			);
+		}
 	}
 
 }
