@@ -4,8 +4,8 @@ namespace FastyBird\Connector\NsPanel\Tests\Cases\Unit;
 
 use Error;
 use FastyBird\Connector\NsPanel\DI;
-use FastyBird\Library\Bootstrap\Boot as BootstrapBoot;
-use FastyBird\Library\Bootstrap\Exceptions as BootstrapExceptions;
+use FastyBird\Library\Application\Boot as ApplicationBoot;
+use FastyBird\Library\Application\Exceptions as ApplicationExceptions;
 use Nette;
 use PHPUnit\Framework\TestCase;
 use function constant;
@@ -13,7 +13,6 @@ use function defined;
 use function file_exists;
 use function getmypid;
 use function md5;
-use function strval;
 use function time;
 
 abstract class BaseTestCase extends TestCase
@@ -22,7 +21,7 @@ abstract class BaseTestCase extends TestCase
 	protected Nette\DI\Container $container;
 
 	/**
-	 * @throws BootstrapExceptions\InvalidArgument
+	 * @throws ApplicationExceptions\InvalidArgument
 	 * @throws Error
 	 */
 	protected function setUp(): void
@@ -33,7 +32,7 @@ abstract class BaseTestCase extends TestCase
 	}
 
 	/**
-	 * @throws BootstrapExceptions\InvalidArgument
+	 * @throws ApplicationExceptions\InvalidArgument
 	 * @throws Error
 	 */
 	protected function createContainer(string|null $additionalConfig = null): Nette\DI\Container
@@ -41,12 +40,12 @@ abstract class BaseTestCase extends TestCase
 		$rootDir = __DIR__ . '/../..';
 		$vendorDir = defined('FB_VENDOR_DIR') ? constant('FB_VENDOR_DIR') : $rootDir . '/../vendor';
 
-		$config = BootstrapBoot\Bootstrap::boot();
+		$config = ApplicationBoot\Bootstrap::boot();
 		$config->setForceReloadContainer();
 		$config->setTempDirectory(FB_TEMP_DIR);
 
 		$config->addStaticParameters(
-			['container' => ['class' => 'SystemContainer_' . strval(getmypid()) . md5((string) time())]],
+			['container' => ['class' => 'SystemContainer_' . getmypid() . md5((string) time())]],
 		);
 		$config->addStaticParameters(['appDir' => $rootDir, 'wwwDir' => $rootDir, 'vendorDir' => $vendorDir]);
 
