@@ -4,19 +4,18 @@ namespace FastyBird\Connector\NsPanel\Tests\Cases\Unit\Clients;
 
 use Error;
 use FastyBird\Connector\NsPanel\Clients;
-use FastyBird\Connector\NsPanel\Entities;
+use FastyBird\Connector\NsPanel\Documents;
 use FastyBird\Connector\NsPanel\Exceptions;
 use FastyBird\Connector\NsPanel\Helpers;
+use FastyBird\Connector\NsPanel\Queries;
 use FastyBird\Connector\NsPanel\Queue;
 use FastyBird\Connector\NsPanel\Services;
 use FastyBird\Connector\NsPanel\Tests;
 use FastyBird\Connector\NsPanel\Types;
-use FastyBird\Library\Bootstrap\Exceptions as BootstrapExceptions;
-use FastyBird\Library\Metadata\Documents as MetadataDocuments;
+use FastyBird\Library\Application\Exceptions as ApplicationExceptions;
 use FastyBird\Library\Metadata\Exceptions as MetadataExceptions;
 use FastyBird\Module\Devices\Exceptions as DevicesExceptions;
 use FastyBird\Module\Devices\Models as DevicesModels;
-use FastyBird\Module\Devices\Queries as DevicesQueries;
 use Nette\DI;
 use Nette\Utils;
 use Psr\Http;
@@ -24,11 +23,15 @@ use React;
 use React\EventLoop;
 use RuntimeException;
 
+/**
+ * @runTestsInSeparateProcesses
+ * @preserveGlobalState disabled
+ */
 final class DiscoveryTest extends Tests\Cases\Unit\DbTestCase
 {
 
 	/**
-	 * @throws BootstrapExceptions\InvalidArgument
+	 * @throws ApplicationExceptions\InvalidArgument
 	 * @throws DevicesExceptions\InvalidState
 	 * @throws DI\MissingServiceException
 	 * @throws Exceptions\InvalidArgument
@@ -81,79 +84,18 @@ final class DiscoveryTest extends Tests\Cases\Unit\DbTestCase
 			DevicesModels\Configuration\Connectors\Repository::class,
 		);
 
-		$findConnectorQuery = new DevicesQueries\Configuration\FindConnectors();
+		$findConnectorQuery = new Queries\Configuration\FindConnectors();
 		$findConnectorQuery->byIdentifier('ns-panel');
-		$findConnectorQuery->byType(Entities\NsPanelConnector::TYPE);
 
-		$connector = $connectorsConfigurationRepository->findOneBy($findConnectorQuery);
-		self::assertInstanceOf(MetadataDocuments\DevicesModule\Connector::class, $connector);
+		$connector = $connectorsConfigurationRepository->findOneBy(
+			$findConnectorQuery,
+			Documents\Connectors\Connector::class,
+		);
+		self::assertInstanceOf(Documents\Connectors\Connector::class, $connector);
 
 		$clientFactory = $this->getContainer()->getByType(Clients\DiscoveryFactory::class);
 
 		$client = $clientFactory->create($connector);
-
-		$client->on('finished', static function (array $foundSubDevices): void {
-			self::assertCount(1, $foundSubDevices);
-
-			$data = [];
-
-			foreach ($foundSubDevices as $gatewaySubDevices) {
-				self::assertCount(1, $gatewaySubDevices);
-
-				foreach ($gatewaySubDevices as $subDevice) {
-					self::assertInstanceOf(Entities\Clients\DiscoveredSubDevice::class, $subDevice);
-
-					$data[] = $subDevice->toArray();
-				}
-			}
-
-			self::assertSame(
-				[
-					[
-						'serial_number' => 'a480062416',
-						'third_serial_number' => null,
-						'service_address' => null,
-						'name' => 'Temperature/Humidity Sensor',
-						'manufacturer' => 'eWeLink',
-						'model' => 'TH01',
-						'firmware_version' => '0.5',
-						'hostname' => null,
-						'mac_address' => '00124b002a5d75b1',
-						'app_name' => null,
-						'display_category' => 'temperatureAndHumiditySensor',
-						'capabilities' => [
-							[
-								'capability' => 'temperature',
-								'permission' => 'read',
-								'name' => null,
-							],
-							[
-								'capability' => 'humidity',
-								'permission' => 'read',
-								'name' => null,
-							],
-							[
-								'capability' => 'battery',
-								'permission' => 'read',
-								'name' => null,
-							],
-							[
-								'capability' => 'rssi',
-								'permission' => 'read',
-								'name' => null,
-							],
-						],
-						'protocol' => 'zigbee',
-						'tags' => [
-							'temperature_unit' => 'c',
-						],
-						'online' => true,
-						'subnet' => true,
-					],
-				],
-				$data,
-			);
-		});
 
 		$client->discover();
 
@@ -177,19 +119,21 @@ final class DiscoveryTest extends Tests\Cases\Unit\DbTestCase
 			DevicesModels\Configuration\Devices\Repository::class,
 		);
 
-		$findDeviceQuery = new DevicesQueries\Configuration\FindDevices();
+		$findDeviceQuery = new Queries\Configuration\FindSubDevices();
 		$findDeviceQuery->forConnector($connector);
 		$findDeviceQuery->byIdentifier('a480062416');
-		$findDeviceQuery->byType(Entities\Devices\SubDevice::TYPE);
 
-		$device = $devicesConfigurationRepository->findOneBy($findDeviceQuery);
+		$device = $devicesConfigurationRepository->findOneBy(
+			$findDeviceQuery,
+			Documents\Devices\SubDevice::class,
+		);
 
 		$deviceHelper = $this->getContainer()->getByType(Helpers\Devices\SubDevice::class);
 
-		self::assertInstanceOf(MetadataDocuments\DevicesModule\Device::class, $device);
+		self::assertInstanceOf(Documents\Devices\SubDevice::class, $device);
 		self::assertSame(
 			Types\Category::TEMPERATURE_HUMIDITY_SENSOR,
-			$deviceHelper->getDisplayCategory($device)->getValue(),
+			$deviceHelper->getDisplayCategory($device),
 		);
 		self::assertSame('eWeLink', $deviceHelper->getManufacturer($device));
 		self::assertSame('TH01', $deviceHelper->getModel($device));
@@ -198,19 +142,21 @@ final class DiscoveryTest extends Tests\Cases\Unit\DbTestCase
 			DevicesModels\Configuration\Channels\Repository::class,
 		);
 
-		$findChannelsQuery = new DevicesQueries\Configuration\FindChannels();
+		$findChannelsQuery = new Queries\Configuration\FindChannels();
 		$findChannelsQuery->forDevice($device);
-		$findChannelsQuery->byType(Entities\NsPanelChannel::TYPE);
 
-		$channels = $channelsConfigurationRepository->findAllBy($findChannelsQuery);
+		$channels = $channelsConfigurationRepository->findAllBy(
+			$findChannelsQuery,
+			Documents\Channels\Channel::class,
+		);
 
-		$channelHelper = $this->getContainer()->getByType(Helpers\Channel::class);
+		$channelHelper = $this->getContainer()->getByType(Helpers\Channels\Channel::class);
 
 		self::assertCount(4, $channels);
 
 		foreach ($channels as $channel) {
 			self::assertContains(
-				$channelHelper->getCapability($channel)->getValue(),
+				$channelHelper->getCapability($channel),
 				[
 					Types\Capability::TEMPERATURE,
 					Types\Capability::HUMIDITY,
