@@ -3,10 +3,10 @@
 namespace FastyBird\Connector\NsPanel\Tests\Cases\Unit\Entities\API;
 
 use Error;
-use FastyBird\Connector\NsPanel\Entities;
+use FastyBird\Connector\NsPanel\API\Messages\Header;
 use FastyBird\Connector\NsPanel\Tests;
 use FastyBird\Connector\NsPanel\Types;
-use FastyBird\Library\Bootstrap\Exceptions as BootstrapExceptions;
+use FastyBird\Library\Application\Exceptions as ApplicationExceptions;
 use Nette;
 use Orisai\ObjectMapper;
 use Ramsey\Uuid;
@@ -16,12 +16,12 @@ final class HeaderTest extends Tests\Cases\Unit\BaseTestCase
 {
 
 	/**
-	 * @throws BootstrapExceptions\InvalidArgument
+	 * @throws ApplicationExceptions\InvalidArgument
 	 * @throws Nette\DI\MissingServiceException
 	 * @throws ObjectMapper\Exception\InvalidData
 	 * @throws Error
 	 */
-	public function testCreateEntity(): void
+	public function testCreateMessage(): void
 	{
 		$container = $this->createContainer();
 
@@ -30,18 +30,18 @@ final class HeaderTest extends Tests\Cases\Unit\BaseTestCase
 
 		$id = Uuid\Uuid::uuid4();
 
-		$entity = $processor->process(
+		$message = $processor->process(
 			[
-				'name' => Types\Header::ERROR_RESPONSE,
+				'name' => Types\Header::ERROR_RESPONSE->value,
 				'message_id' => $id->toString(),
 				'version' => '1',
 			],
-			Entities\API\Header::class,
+			Header::class,
 		);
 
-		self::assertSame(Types\Header::ERROR_RESPONSE, $entity->getName()->getValue());
-		self::assertSame($id->toString(), $entity->getMessageId());
-		self::assertSame('1', $entity->getVersion());
+		self::assertSame(Types\Header::ERROR_RESPONSE, $message->getName());
+		self::assertSame($id->toString(), $message->getMessageId());
+		self::assertSame('1', $message->getVersion());
 	}
 
 }
